% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_mc_metadata}
\alias{add_mc_metadata}
\title{Add per-metacell metadata to an McATAC object}
\usage{
add_mc_metadata(mcatac, metadata)
}
\arguments{
\item{mcatac}{an McATAC object}

\item{metadata}{data frame with a column called 'metacell' and additional metacell annotations, or the name of a delimited file which contains such annotations.}
}
\description{
Add per-metacell metadata to an McATAC object
}
\examples{
\dontrun{
data(mcmd)
mc_atac <- add_metadata(mc_atac, mcmd)
}

}
