% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{add_mc_rna}
\alias{add_mc_rna}
\title{Add per-metacell gene expression data to an McPeaks object}
\usage{
add_mc_rna(atac_mc, mc_rna)
}
\arguments{
\item{atac_mc}{an McPeaks/McTracks object}

\item{mc_rna}{a \code{metacell1} 'mc' object or a \code{metacells} metacell UMI matrix (a matrix where each row is a gene and each column is a metacell)}
}
\description{
Add per-metacell gene expression data to an McPeaks object
}
\examples{
\dontrun{
data(mcmd)
metacell::scdb_init("pbmc_data/scdb", force_reinit = TRUE)
mc_rna <- metacell::scdb_mc("rna")
atac_mc <- add_mc_rna(atac_mc, mc_rna)
}

}
