% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_peaks}
\alias{annotate_peaks}
\title{Annotate ATAC peaks}
\usage{
annotate_peaks(tss, exons)

annotate_peaks(tss, exons)
}
\arguments{
\item{tss}{tss interval set built-in to misha}

\item{exons}{exon interval set built-in to misha}

\item{atac}{a McATAC or ScATAC object}
}
\value{
the original \code{atac} object, with the peaks annotated using \code{annotate_intervals}

gene_body_df a misha intervals set approximating starts and ends of genes
}
\description{
Annotate ATAC peaks

Generate dataframe of approximate gene bodies (edges of first and last exons)
}
