% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\name{atac_ignore_cells}
\alias{atac_ignore_cells}
\title{Remove cells}
\usage{
atac_ignore_cells(atac, ig_cells)
}
\arguments{
\item{atac}{an ScATAC object}

\item{ig_cells}{a vector of cell names (10X barcodes) to removes}
}
\value{
an scATAC object with the given cells removed.
}
\description{
This function removes a given list of cells from the {.code mat} slot.
}
\examples{
\dontrun{
cs <- Matrix::colSums(atac_sc@mat)
big_cells <- names(cs)[which(cs >= quantile(cs, 0.98))]
atac_sc_filtered <- atac_ignore_cells(atac_sc, big_cells)
}
}
