% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\name{atac_ignore_peaks}
\alias{atac_ignore_peaks}
\title{Set ignored (i.e. blacklisted) peaks}
\usage{
atac_ignore_peaks(atac, ig_peaks)
}
\arguments{
\item{atac}{an ScATAC or McATAC object}

\item{ig_peaks}{a PeakIntervals object, or vector of peak names to ignore}
}
\value{

}
\description{
Given a list of peaks to ignore, this will cancel any previous policy for blacklisting and remove the given peaks from the {.code ignore_peaks} and {.code ignore_pmat} slots. Note that the matrix and peaks would be reordered.
}
\examples{
\dontrun{

}
}
