% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\docType{class}
\name{ATACPeaks-class}
\alias{ATACPeaks-class}
\alias{ATACPeaks}
\title{ATACPeaks objects}
\description{
ATACPeaks is a shallow object holding ATAC data over cells/metacells. Minimally it should include a count matrix of peaks over cells/metacells, \code{PeakIntervals} which hold the coordinates of the peaks and the id of the genome assembly of the peaks. ScATACPeaks and
McATACPeaks extend the ATACPeaks object by adding metadata and additional slots.
}
\section{Slots}{

\describe{
\item{\code{id}}{an identifier for the object, e.g. "pbmc".}

\item{\code{description}}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"}

\item{\code{mat}}{a numeric matrix where rows are peaks and columns are cells/metacells. Can be a sparse matrix.}

\item{\code{peaks}}{misha intervals set. Can contain a field named 'peak_name' with a unique name per peak. Both the names and intervals should be unique (a peak cannot appear more than once).}

\item{\code{genome}}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{\code{metadata}}{data frame with a column called 'metacell' and additional metacell annotations for McATACPeaks, or 'cell_id' and per-cell annotatoins for ScATACPeaks. The constructor can also include or the name of a delimited file which contains such annotations.}

\item{\code{path}}{original path from which the object was loaded (optional)}

\item{\code{promoters}}{are the peaks promoters (optional). When peaks are promoters the peak name would be the gene name instead of coordinates.}

\item{\code{tad_based}}{are the peaks named based on TADs.}
}}

