% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-raw-counts.R
\name{bam_file_chromosomes}
\alias{bam_file_chromosomes}
\title{Get chromosomes from a bam file}
\usage{
bam_file_chromosomes(bam_file, samtools_bin = "samtools")
}
\arguments{
\item{bam_file}{name of the bam file}

\item{samtools_bin}{path to samtools executable}
}
\value{
a vector of strings with the chromosome names
}
\description{
Get chromosomes from a bam file
}
\examples{
\dontrun{
bam_file_chromosomes("pbmc_data/possorted_bam.bam")
}

}
