% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bam.r
\name{bam_to_wig}
\alias{bam_to_wig}
\title{Convert BAM file to WIG file}
\usage{
bam_to_wig(
  bam_path,
  output_filename,
  track_name_prefix = NULL,
  header_line = NULL
)
}
\arguments{
\item{bam_path}{path to the BAM file}

\item{output_filename}{what to call the output WIG file}

\item{track_name_prefix}{(optional) name to prepend to track (e.g. for display in UCSC genome browser)}
}
\description{
Convert BAM file to WIG file
}
