% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr-bias.R
\name{calc_const_peaks}
\alias{calc_const_peaks}
\title{Calculate constitutive peaks}
\usage{
calc_const_peaks(mcatac, const_threshold)
}
\arguments{
\item{mcatac}{McPeaks object.}

\item{const_threshold}{The threshold value for determining constitutive peaks. A peak would be considered constitutive if its minimal expression across all cell types is greater than or equal to this threshold.}
}
\value{
A named logical vector indicating whether each peak is constitutive or not.
}
\description{
This function calculates the constitutive peaks based on the minimal expression of each peak across different cell types.
}
\examples{
\dontrun{
mcatac <- import_from_matrix(mat, peaks, genome = "mm10", class = "McPeaks")
mcatac <- normalize_egc(mcatac, "gastrulation.marginal")
const_v <- calc_const_peaks(mcatac, -16)
}

}
