% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoters.R
\name{calc_prom_rna_cor}
\alias{calc_prom_rna_cor}
\title{Calculate the cross-correlation between promoter accessibility and RNA expression}
\usage{
calc_prom_rna_cor(
  atac_mc,
  rm_zeros = TRUE,
  match_genes = FALSE,
  spearman = FALSE,
  pairwise.complete.obs = TRUE
)
}
\arguments{
\item{atac_mc}{a McATAC object with promoters (using \code{gen_promoter_features}) and RNA expression (using \code{add_mc_rna})}

\item{rm_zeros}{remove genes with no RNA expression in any metacell. Default: TRUE}

\item{match_genes}{calculate correlation only between genes and promoters which have both accesability and RNA expression.
Matching is done based on name. Default: FALSE}

\item{spearman}{if 'TRUE' Spearman correlation is computed, otherwise
Pearson}

\item{pairwise.complete.obs}{see below}
}
\value{
a correlation matrix where rows are promoters and columns are genes
}
\description{
Calculate the cross-correlation between promoter accessibility and RNA expression
}
