% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{calculate_d_stats}
\alias{calculate_d_stats}
\title{Calculate Kolmogorov-Smirnov D statistics between two interval sets with motif energies}
\usage{
calculate_d_stats(pssm_fg, pssm_bg, fg_clustering = NULL, parallel = TRUE)
}
\arguments{
\item{pssm_fg}{motif energies calculated for a certain set of motifs on a PeakIntervals/ScATAC/McATAC object}

\item{pssm_bg}{a background set of intervals (e.g. random genome, all ENCODE enhancers etc.) that all/subset of the motifs (columns) in pssm_fg}

\item{fg_clustering}{a vector of cluster assignments for the foreground peaks (e.g.)}

\item{parallel}{(optional) - whether to use parallelize computations}
}
\value{
a matrix of peaks (rows) vs. aggregated motif energies (columns)
}
\description{
Calculate Kolmogorov-Smirnov D statistics between two interval sets with motif energies
}
\examples{
\dontrun{
     d_vs_rg <- calculate_d_stats(pssm_fg = my_peak_motif_mat, pssm_bg <- random_genome_motif_mat,
                               parallel = F)
}
}
