% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{call_peaks}
\alias{call_peaks}
\title{Call peaks from ATAC marginals track}
\usage{
call_peaks(
  marginal_track,
  quantile_thresh = 0.95,
  min_umis = 8,
  max_peak_size = 600,
  min_peak_size = 50,
  genome = NULL
)
}
\arguments{
\item{marginal_track}{Name of the 'misha' track to call peaks from. You can create it using {.code import_atac_marginal}.}

\item{quantile_thresh}{Quantile of the marginal track above which peaks are called.}

\item{min_umis}{Minimal number of UMIs to call a peak.}

\item{max_peak_size}{Maximal size of a peak. Peaks above this size would be splitted equally into smaller peaks.}

\item{min_peak_size}{Minimal size of a peak.}

\item{genome}{Genome name, such as 'hg19' or 'mm10'. If NULL - the current misha database is used.}
}
\value{
an intervals set with the called peaks
}
\description{
Peaks are called by screening for genomic regions with a number of UMIs above a quantile of the marginal ATAC counts,
and above \code{min_umis}. Peaks that are longer than \code{max_peak_size} would be splitted equally into smaller peaks.
}
\examples{
\dontrun{
peaks <- call_peaks("pbmc_atac.marginal", quantile_thresh = 0.95, min_umis = 8, max_peak_size = 600, genome = "hg38")
}

}
