% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Test if mcATAC has the required dependencies}
\usage{
check_dependencies()
}
\value{
a vector of the dependencies that are not available (invisibly)
}
\description{
Tests if mcATAC has the following dependencies: "grep", "awk", "zcat", "sed", "sort", "head", "tail", "wc" and "uniq". If samtools or tabix are not installed, a warning is issued.
}
\examples{
check_dependencies()

}
