% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{create_bundle}
\alias{create_bundle}
\title{Create a bundle for the shiny app in order to run it with shiny server}
\usage{
create_bundle(
  mct,
  path,
  overwrite = FALSE,
  self_contained = FALSE,
  branch = "master",
  restart = overwrite,
  permissions = NULL,
  ...
)
}
\arguments{
\item{mct}{a MCTracks object}

\item{path}{Path to the bundle directory}

\item{overwrite}{overwrite bundle if already exists}

\item{self_contained}{include the source code of \code{mcATAC} in the bundle
and use it to run the app. Use this in order to ensure that the package would always
run the same way, regardless of mcATAC changes. When this option is FALSE,
the installed version of \code{mcATAC} would be loaded, which can be occasionally
updated for all the \code{mcATAC} apps running from a server. By default, the code
of the latest \code{mcATAC} release would be used, see \code{branch} for
other options.}

\item{branch}{name of the \code{mcATAC} branch to include when \code{self_contained=TRUE}. By default, the master branch would be used.}

\item{restart}{add a file named 'restart.txt' to the bundle. This would force shiny-server to restart the app when updated.}

\item{permissions}{change the file permissions of the bundle after creation, e.g. "777". When NULL -
permissions would not be changed.}

\item{...}{
  Arguments passed on to \code{\link[gert:git_fetch]{gert::git_clone}}
  \describe{
    \item{\code{password}}{a string or a callback function to get passwords for authentication
or password protected ssh keys. Defaults to \link[askpass:askpass]{askpass} which
checks \code{getOption('askpass')}.}
    \item{\code{ssh_key}}{path or object containing your ssh private key. By default we
look for keys in \code{ssh-agent} and \link[credentials:ssh_credentials]{credentials::ssh_key_info}.}
    \item{\code{verbose}}{display some progress info while downloading}
    \item{\code{mirror}}{use the \code{--mirror} flag}
    \item{\code{url}}{remote url. Typically starts with \verb{https://github.com/} for public
repositories, and \verb{https://yourname@github.com/} or \verb{git@github.com/} for
private repos. You will be prompted for a password or pat when needed.}
    \item{\code{bare}}{use the \code{--bare} flag}
  }}
}
\description{
Generate a 'deployment ready' bundle of the shiny app
}
\details{
Create a minimal shiny app in \code{path} directory which would contain:
\itemize{
\item{}{app.R file. }
\item{}{mct.rds with the mct object. }
}

The bundle can then be deployed in shiny-server, shinyapps.io or any other environment that supports serving shiny apps.

Note: when deploying to these services - make sure you have the mcATAC package installed with the 'Suggests' dependencies.
}
\examples{
\dontrun{
create_bundle(mct, "/path/to/the/bundle/directory")
}

}
