% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_atac_clust_ucsc}
\alias{export_atac_clust_ucsc}
\title{Generate a ucsc genome browser file for each metacell cluster}
\usage{
export_atac_clust_ucsc(mc_atac, mc_atac_clust, fn)
}
\arguments{
\item{mc_atac}{McATAC object}

\item{mc_atac_clust}{output of \code{gen_atac_mc_clust}}

\item{fn}{name of the genome browser file (should have a '.ucsc' extension)}
}
\description{
Generate a ucsc genome browser file for each metacell cluster
}
