% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.r
\name{find_blacklist_overlaps}
\alias{find_blacklist_overlaps}
\title{Find overlaps with ENCODE blacklists}
\usage{
find_blacklist_overlaps(
  scatac = NULL,
  mcatac = NULL,
  peaks = NULL,
  genome = NULL
)
}
\arguments{
\item{scatac}{(optional) an ScATAC object}

\item{mcatac}{(optional) an McATAC object}

\item{peaks}{(optional) the intervals set to check}

\item{genome}{(optional, required if checking peaks directly) the interval set to check}
}
\value{
blacklist_overlaps - a PeakIntervals object with peaks identified as overlapping blacklisted regions
}
\description{
Identify peaks in the data which overlap (or are adjacent to?) regions blacklisted by ENCODE as having universally high DNAse HS or ChIP signal (basically mapping artifacts)
}
\examples{
\dontrun{
my_intervals <- annotate_intervals(my_intervals, "mm10", min_proximal = 1e+03, max_proximal = 2e+04, max_distal = 1e+06, exonic_peak_dist = 5e+2)
table(my_intervals$peak_annot)
}
}
