% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.r
\name{fix_missing_chroms_in_peaks}
\alias{fix_missing_chroms_in_peaks}
\title{Utility function that makes fake intervals for chromosome/contig not present in an interval set
Explanation: as a validation step, gextract requires that every chromosome/contig in the ALLGENOME
variable also be present in the track directory, so we add fake sequences in all
missing chromosomes/contigs}
\usage{
fix_missing_chroms_in_peaks(peaks)
}
\description{
Utility function that makes fake intervals for chromosome/contig not present in an interval set
Explanation: as a validation step, gextract requires that every chromosome/contig in the ALLGENOME
variable also be present in the track directory, so we add fake sequences in all
missing chromosomes/contigs
}
