% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakIntervals.R
\name{gen_peak_features}
\alias{gen_peak_features}
\title{Select peaks features by minimal coverage and threshold max-min fold}
\usage{
gen_peak_features(
  atac,
  min_fold,
  min_abs,
  only_promoters = FALSE,
  upstream = 500,
  downstream = 50
)
}
\arguments{
\item{atac}{a McATAC or ScATAC object}

\item{min_fold}{minimal max-min fold change}

\item{min_abs}{minimal total coverage per peak}

\item{only_promoters}{when TRUE - summarise over features to get promoters}

\item{upstream}{bp upstream to TSS}

\item{downstream}{bp downstread from tss}
}
\description{
Select peaks features by minimal coverage and threshold max-min fold
}
