% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoters.R
\name{gen_promoter_features}
\alias{gen_promoter_features}
\title{Select only peaks that are close to a gene promoter}
\usage{
gen_promoter_features(
  atac,
  upstream = 500,
  downstream = 50,
  tss_intervals = "intervs.global.tss",
  id_field = "geneSymbol"
)
}
\arguments{
\item{atac}{a McATAC or ScATAC object}

\item{upstream}{bp upstream to TSS}

\item{downstream}{bp downstread from tss}

\item{tss_intervals}{name of the intervals set with the TSS}

\item{id_field}{the field in \code{tss_intervals} containing the gene names. Default: "geneSymbol"}
}
\value{
the McATAC or ScATAC object with new peaks which are a sum of the peaks within the defined promoters. The rownames of the \code{atac@mat} and the field 'peak_name' in \code{atac@peaks} would contain the name of the promoter (from \code{id_field}).
Note that previous peak metadata and previous ignored peaks are dropped.
}
\description{
Select only peaks that are close to a gene promoter. In case of alternative promoters or multiple peaks close to the same promoter, the reads from all the peaks and promoters are summed.
}
\examples{
\dontrun{
atac_mc_promoters <- gen_promoter_features(atac_mc)
}

}
