% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.r
\name{generate_motif_pssm_matrix}
\alias{generate_motif_pssm_matrix}
\title{Generate peak-motif matrix}
\usage{
generate_motif_pssm_matrix(
  atac = NULL,
  peak_set = NULL,
  peak_width = 200,
  pssm_path = NULL,
  datasets_of_interest = NULL,
  motif_tracks = NULL,
  motif_regex = NULL,
  parallel = FALSE
)
}
\arguments{
\item{atac}{(optional) - an ScATAC/McATAC object}

\item{peak_set}{(optional) - a PeakIntervals object}

\item{peak_width}{(optional) - size of region around peak centers to extract motif energies for}

\item{pssm_path}{(optional) - path to directory containing misha-formatted pssm files (e.g. \code{motifs.key} and \code{motifs.data})}

\item{datasets_of_interest}{(optional) - names of pssm datasets (\link{name}.key-\link{name}.data file combinations) to calculate PSSM values for}

\item{motif_tracks}{(optional) - misha track names for which to extract motif PSSMs}

\item{motif_regex}{(optional) - a vector of regular expressions for which to match motif track names and extract motif PSSMs}

\item{parallel}{(optional) - whether to use parallel computations}
}
\value{
a matrix of peaks (rows) vs. aggregated motif energies (columns)
}
\description{
This function calculates aggregated log-PSSM energies (as derived by the misha package's \code{create_pssm_energy} module) for a set of peaks
(default - all peaks in the dataset) and for a set of motifs (default - all available)
}
\examples{
\dontrun{
     peak_motif_mat <- generate_motif_pssm_matrix(peak_set=head(scatac@peaks), 
                               motif_regex = c("Bcl", "Atf"), 
                               datasets_of_interest = c("homer", "jaspar", "jolma"),
                               parallel = F)
}
}
