% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bam.R
\name{generate_per_metacell_bams}
\alias{generate_per_metacell_bams}
\title{Subset 10X atac_possorted.bam file into per-metacell BAM files}
\usage{
generate_per_metacell_bams(
  bam_path,
  mcatac,
  out_dir = file.path(bam_path, paste0(mcatac@id, "_mc_bams")),
  gparallel_path = "/usr/wisdom/parallel",
  samtools_path = "samtools"
)
}
\arguments{
\item{bam_path}{path to the 10X atac_possorted.bam file}

\item{mcatac}{an McATAC object (from which to get metacell assignments)}

\item{out_dir}{(optional) directory to output per-metacell bam files. Default: /bam_path/mcatac@id_mc_bams}

\item{gparallel_path}{(optional) path to GNU parallel}

\item{samtools_path}{(optional) path to samtools (requires a version which supports the -D/--tag-file option in samtools view)}
}
\value{
error log
}
\description{
Subset 10X atac_possorted.bam file into per-metacell BAM files
}
