% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bam.r
\name{generate_wigs_from_bams}
\alias{generate_wigs_from_bams}
\title{Write WIGs from BAMs}
\usage{
generate_wigs_from_bams(
  bam_folder_path,
  track_name_prefix,
  output_path = NULL,
  parallel = TRUE
)
}
\arguments{
\item{bam_folder_path}{path to BAMs}

\item{track_name_prefix}{(optional) name to prepend to track (e.g. for display in UCSC genome browser)}

\item{output_path}{(optional) where to put WIGs}

\item{parallel}{(optional) whether to do parallel computations}
}
\description{
Write WIGs from BAMs
}
