% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.r
\name{get_available_pssms}
\alias{get_available_pssms}
\title{Utility function for validating \code{pssm_path} and getting all PSSM tracks from path/gdb}
\usage{
get_available_pssms(pssm_path = NULL, datasets_of_interest = NULL)
}
\description{
Utility function for validating \code{pssm_path} and getting all PSSM tracks from path/gdb
}
