% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{get_gene_body_df}
\alias{get_gene_body_df}
\title{Generate dataframe of approximate gene bodies (edges of first and last exons) - backend function}
\usage{
get_gene_body_df(tss, exons)
}
\arguments{
\item{tss}{tss interval set built-in to misha}

\item{exons}{exon interval set built-in to misha}
}
\value{
gene_body_df a misha intervals set approximating starts and ends of genes
}
\description{
Generate dataframe of approximate gene bodies (edges of first and last exons) - backend function
}
