% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{get_peak_coverage_stats}
\alias{get_peak_coverage_stats}
\title{Calculate coverage statistics of the peaks}
\usage{
get_peak_coverage_stats(atac, scale = 100)
}
\arguments{
\item{atac}{an ScPeaks or McPeaks object}

\item{scale}{for the 'cov_density' field, calculate the number of UMIs per \code{scale}
bp (default: 100)}
}
\value{
a data frame with peaks, and their length ('len'), their total coverage ('cov'),
maximal per-cell coverage ('max_cov') and their coverage density ('cov_density')
}
\description{
For each peak, calculate:
\enumerate{
\item{cov: }{The total number of UMIs in the peak}
\item{max_cov: }{The maximal per-cell coverage of the peak}
\item{len: }{The length of the peak}
\item{cov_density: }{The number of UMIs the number of UMIs per \code{scale} bp.}
}
}
\examples{
\dontrun{
get_peak_coverage_stats(atac)
}

}
