% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.r
\name{get_peaks_for_pssm}
\alias{get_peaks_for_pssm}
\title{Utility function for validating parameter combinations for PSSM extraction}
\usage{
get_peaks_for_pssm(atac, peak_set)
}
\arguments{
\item{atac}{(optional) - an ScATAC/McATAC object}

\item{peak_set}{(optional) - a PeakIntervals object}
}
\description{
Utility function for validating parameter combinations for PSSM extraction
}
