% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakIntervals.R
\name{get_promoter_peaks}
\alias{get_promoter_peaks}
\title{Get the names of the peaks close to a gene promoter}
\usage{
get_promoter_peaks(
  peaks,
  gene,
  max_dist_to_promoter_peak = 500,
  tss_intervals = "intervs.global.tss"
)
}
\arguments{
\item{peaks}{a PeakIntervals object}

\item{gene}{a gene name, e.g. "CD4"}

\item{max_dist_to_promoter_peak}{how far from \code{gene}'s TSS to search for promoter-proximal peaks. Default: 500}

\item{tss_intervals}{name of the intervals set containing the TSSs}
}
\value{
names of the peaks close to \code{gene}'s promoter. If no peaks are found, NULL is returned.
}
\description{
Get the names of the peaks close to a gene promoter
}
\examples{
\dontrun{
get_promoter_peaks(mc_atac@peaks, "CD4")
}
}
