% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{get_quantile_cov_thresh}
\alias{get_quantile_cov_thresh}
\title{Calculate the threshold for calling peaks}
\usage{
get_quantile_cov_thresh(
  marginal_track,
  quantile_thresh,
  min_umis,
  genome = NULL,
  seed = 60427
)
}
\arguments{
\item{marginal_track}{Name of the 'misha' track to call peaks from. You can create it using \code{import_atac_marginal}.}

\item{quantile_thresh}{Quantile threshold to use.}

\item{min_umis}{Minimum number of UMIs to use.}

\item{genome}{Genome name, such as 'hg19' or 'mm10'. If NULL - the current genome is used.}

\item{seed}{random seed for reproducibility (\code{misha::gquantiles} sometimes samples the data.}
}
\value{
The number of UMIs to use for calling peaks.
}
\description{
Threshold is calculated by taking the maximum between the \code{quantile_thresh} quantile of the ATAC marginal and \code{min_umis}.
}
\examples{
\dontrun{
get_quantile_cov_thresh("pbmc_atac.marginal", quantile_thresh = 0.9, min_umis = 8, genome = "hg38")
}

}
