% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{get_rna_egc}
\alias{get_rna_egc}
\title{Get normalized counts of gene expression}
\usage{
get_rna_egc(atac_mc, genes = NULL, rm_zeros = TRUE, epsilon = 0.00001)
}
\arguments{
\item{atac_mc}{a McATAC object with RNA expression (using \code{add_mc_rna})}

\item{genes}{list of genes to match. Default (NULL): all genes}

\item{rm_zeros}{remove genes with no RNA expression in any metacell. Default: TRUE}

\item{epsilon}{regularization factor added to the log normalized expression}
}
\value{
a matrix with normalized counts of gene expression for each gene (rows) and metacell (columns)
}
\description{
Get counts of a gene expression normalized by sum of counts per metacell.
}
\examples{
\dontrun{
get_rna_egc(atac_mc)
get_rna_egc(atac_mc, "GNLY")
get_rna_egc(atac_mc, "GNLY", epsilon = NULL)
}

}
