% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{get_rna_matrix}
\alias{get_rna_matrix}
\title{Get the RNA expression matrix from a McATAC object}
\usage{
get_rna_matrix(atac_mc, genes = NULL, rm_zeros = TRUE)
}
\arguments{
\item{atac_mc}{a McATAC object with RNA expression (using \code{add_mc_rna})}

\item{genes}{list of genes to match. Default (NULL): all genes}

\item{rm_zeros}{remove genes with no RNA expression in any metacell. Default: TRUE}
}
\value{
a matrix with RNA expression values for each gene (rows) and metacell (columns)
}
\description{
Get the RNA expression matrix from a McATAC object
}
\examples{
\dontrun{
rna_mat <- get_rna_matrix(atac_mc)
rna_mat <- get_rna_matrix(atac_mc, genes = c("MESP1", "MESP2", "PF4"))
rna_mat <- get_rna_matrix(atac_mc, rm_zeros = FALSE)
}

}
