% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_track_total_coverage}
\alias{get_track_total_coverage}
\title{Return the total coverage of a track}
\usage{
get_track_total_coverage(track, attr_name = "total_cov", overwrite = FALSE)
}
\arguments{
\item{track}{track name}

\item{attr_name}{attribute name. Default is "total_cov", please do not change this unless you know what you are doing.}

\item{overwrite}{overwrite existing attribute (logical)}
}
\value{
Total coverage of the track.
}
\description{
This function computes the total coverage of a track, and
adds it as an attribute to the track. If the attribute was pre-computed
before, it is returned. Set the \code{overwrite} parameter to TRUE to recompute the coverage.
}
\examples{
\dontrun{
add_track_total_coverage("track1")
}

}
