% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-intervals.R
\name{gintervals.extend_left}
\alias{gintervals.extend_left}
\title{Extend the coordinates of an intervals set to the left.}
\usage{
gintervals.extend_left(intervals, ext)
}
\arguments{
\item{intervals}{An intervals set.}

\item{ext}{number of bp to extend to the left.}
}
\value{
A new intervals set where the start coordinates are \code{start - ext}. Note that the intervals remain bounded by the chromosome boundaries.
}
\description{
Extend the coordinates of an intervals set to the left.
}
\examples{
gdb.init_examples()
gintervals(1, 50, 150) \%>\% gintervals.extend_left(20)

}
