% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-intervals.R
\name{gintervals.extend_right}
\alias{gintervals.extend_right}
\title{Extend the coordinates of an intervals set to the right.}
\usage{
gintervals.extend_right(intervals, ext)
}
\arguments{
\item{intervals}{An intervals set.}

\item{ext}{number of bp to extend to the right.}
}
\value{
A new intervals set where the end coordinates are \code{end + ext}. Note that the intervals remain bounded by the chromosome boundaries.
}
\description{
Extend the coordinates of an intervals set to the right.
}
\examples{
gdb.init_examples()
gintervals(1, 50, 150) \%>\% gintervals.extend_right(20)

}
