% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-intervals.R
\name{gintervals.shift_left}
\alias{gintervals.shift_left}
\title{Shift the coordinates of an intervals set to the left.}
\usage{
gintervals.shift_left(intervals, shift)
}
\arguments{
\item{intervals}{An intervals set.}

\item{shift}{number of bp to shift to the left.}
}
\value{
A new intervals set where the coordinates are \code{start - shift} and \code{end - shift}. Note that the intervals remain bounded by the chromosome boundaries.
}
\description{
Shift the coordinates of an intervals set to the left.
}
\examples{
gdb.init_examples()
gintervals(1, 50, 150) \%>\% gintervals.shift_left(20)

}
