% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-intervals.R
\name{gintervals.zoom_in}
\alias{gintervals.zoom_in}
\title{Zoom-in for an intervals set.}
\usage{
gintervals.zoom_in(intervals, zoom)
}
\arguments{
\item{intervals}{An intervals set.}

\item{zoom}{the zoom-in factor.}
}
\value{
A new intervals set where the coordinates are \code{c + round(l/2)} and \code{c - round(l/2)} where \code{l} is the length of the interval divided by \code{zoom}, and \code{c} is the center of the interval.
}
\description{
Zoom-in for an intervals set.
}
\examples{
gdb.init_examples()
gintervals(1, 0, 200) \%>\% gintervals.zoom_in(2)

}
