% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-intervals.R
\name{gintervals.zoom_out}
\alias{gintervals.zoom_out}
\title{Zoom-out for an intervals set.}
\usage{
gintervals.zoom_out(intervals, zoom)
}
\arguments{
\item{intervals}{An intervals set.}

\item{zoom}{the zoom-our factor.}
}
\value{
A new intervals set where the coordinates are \code{c + round(l/2)} and \code{c - round(l/2)} where \code{l} is the length of the interval multiplied by \code{zoom}, and \code{c} is the center of the interval. Note that the intervals remain bounded by the chromosome boundaries.
}
\description{
Zoom-out for an intervals set.
}
\examples{
gdb.init_examples()
gintervals(1, 50, 150) \%>\% gintervals.zoom_out(2)

}
