% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-multi-genomes.R
\name{group_to_hclust}
\alias{group_to_hclust}
\title{Generate an hclust object from a data.frame with cell_type, group and order columns}
\usage{
group_to_hclust(df)
}
\arguments{
\item{df}{data.frame with cell_type, group and order}
}
\value{
hclust object with cell_type as labels and order as order of the leaves
}
\description{
Generate an hclust object from a data.frame with cell_type, group and order columns
}
\examples{
df <- data.frame(
    cell_type = c(
        "PGC", "Surface ectoderm", "Neural crest",
        "Neural tube/Floor plate", "Forebrain/Midbrain/Hindbrain", "Caudal neural plate",
        "Rostral neural plate", "Neural plate boundary", "Definitive ectoderm",
        "Epiblast", "Primitive streak", "Caudal epiblast", "Tail bud - neural",
        "Tail bud - mesoderm", "Early nascent mesoderm"
    ), order = 1:15,
    group = c(
        "ecto", "ecto", "ecto", "ecto", "ecto", "ecto",
        "ecto", "ecto", "ecto", "early", "early", "early", "early",
        "Meso", "Meso"
    )
)

group_to_hclust(df)

}
