% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{has_cell_type}
\alias{has_cell_type}
\title{Does the McATACPeaks object contain per-metacell cell type annotation}
\usage{
has_cell_type(atac_mc)
}
\arguments{
\item{atac_mc}{an McATACPeaks object}
}
\value{
TRUE if the McATACPeaks object contains per-metacell cell type annotation, FALSE otherwise
}
\description{
Does the McATACPeaks object contain per-metacell cell type annotation
}
\examples{
\dontrun{
has_cell_type(atac_mc)
}

}
