% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{has_rna}
\alias{has_rna}
\title{Does the McATAC object contain per-metacell gene expression data?}
\usage{
has_rna(mc_atac)
}
\arguments{
\item{mc_atac}{an McATAC object}
}
\value{
TRUE if the McATAC object contains per-metacell gene expression data, FALSE otherwise
}
\description{
Does the McATAC object contain per-metacell gene expression data?
}
\examples{
\dontrun{
has_rna(mc_atac)
}

}
