% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\name{ignore_peaks}
\alias{ignore_peaks}
\title{Set ignored (i.e. blacklisted) peaks}
\usage{
ignore_peaks(obj = NULL, ig_peaks)
}
\arguments{
\item{ig_peaks}{a PeakIntervals object, or vector of peak names to ignore}

\item{mcatac}{an McATAC object}
}
\description{
Given a list of peaks to ignore, this will cancel any previous policy for blacklisting and remove the given genes to the ignore_mat. Downstream algorithm usually ignore these genes altogether, for any purpose including normalization. However, ignored genes can be accessed and analyzed seperately for validation/tests or when they represent some relevant biology (e.g. cell cycle)
}
