% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{import_atac_marginal}
\alias{import_atac_marginal}
\title{Import marginal of ATAC counts to a misha track}
\usage{
import_atac_marginal(
  file,
  track,
  description,
  genome,
  binsize = 20,
  overwrite = FALSE,
  wig_temp_dir = tempdir()
)
}
\arguments{
\item{file}{Name of the 'bigwig' file to import, such as the 'atac_cut_sites.bigwig' from the 10x pipeline.
For more details, see:
https://support.10xgenomics.com/single-cell-multiome-atac-gex/software/pipelines/latest/output/bigwig}

\item{track}{Name of the track to create.}

\item{description}{a character string description}

\item{genome}{Genome name, such as 'hg19' or 'mm10'.}

\item{binsize}{bin size of the newly created 'Dense' track or '0' for a
'Sparse' track}

\item{overwrite}{Overwrite the existing track if it exists.}

\item{wig_temp_dir}{Temporary directory to store the intermediate wig files.}
}
\description{
Import marginal of ATAC counts to a misha track
}
\examples{
\dontrun{
import_atac_marginal("./atac_cut_sites.bigwig", "pbmc_atac.marginal", "Marignal counts from PBMC ATAC", genome = "hg38")
}

}
