% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-peaks.R
\name{import_from_matrix}
\alias{import_from_matrix}
\title{Create an ScATAC,McATAC object from matrix and peaks}
\usage{
import_from_matrix(
  mat,
  peaks,
  genome,
  id = NULL,
  description = NULL,
  metadata = NULL,
  class = "ScATAC",
  rm_zero_peaks = TRUE,
  tad_based = TRUE,
  zero_based = FALSE,
  mc_size_eps_q = 0.1
)
}
\arguments{
\item{mat}{a matrix/sparse matrix of counts}

\item{peaks}{an intervals set where each row corresponds to a row in \code{mat}. Can contain
an additional column named 'peak_name' with peak names, in which case the rownames of \code{mat}
are expected to equal to this column.}

\item{genome}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{id}{an identifier for the object, e.g. "pbmc". If NULL - a random id would be assigned.}

\item{description}{(Optional) description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)".}

\item{metadata}{(Optional) data frame with a column called 'metacell' and additional metacell annotations, or the name of a delimited file which contains such annotations.}

\item{class}{should the output object be a ScATAC or McATAC}

\item{rm_zero_peaks}{remove peaks without any reads (all-zero peaks). Default: TRUE}

\item{tad_based}{whether to name peaks based on TADs if possible (when an intervals set named "intervs.global.tad_names" exists). When FALSE - peaks are named based on their coordinates.}

\item{zero_based}{are the coordinates of the features 0-based? Note that when this is FALSE (default) the coordinates of the created object would be different from the ones in \code{features_fn} by 1 bp.}

\item{mc_size_eps_q}{(optional) quantile of MC size (in UMIs) to scale the number of UMIs per metacell. \eqn{egc_ij} would then be
the fraction of peak i in metacell j multiplied by the \code{mc_size_eps_q} quantile of metacell sizes.}
}
\description{
Create an ScATAC,McATAC object from matrix and peaks
}
\examples{
\dontrun{
atac_sc <- import_from_matrix(mat, peaks, genome = "hg38")
atac_mc <- import_from_matrix(mat, peaks, genome = "hg38", class = "McATAC")
}

}
