% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_track_order}
\alias{infer_track_order}
\title{Infer order of metacell track names by metacell number}
\usage{
infer_track_order(tracks, track_name_prefix = NULL)
}
\arguments{
\item{tracks}{the tracks to order}

\item{track_name_prefix}{prefix string to remove when extracting metacell numbers}
}
\value{
order of tracks by metacell number
}
\description{
Infer order of metacell track names by metacell number
}
\examples{
track_order <- infer_track_order(tracks)
}
