% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATACPeaks.R
\name{initialize,McATACPeaks-method}
\alias{initialize,McATACPeaks-method}
\title{Construct a new McATACPeaks object}
\usage{
\S4method{initialize}{McATACPeaks}(
  .Object,
  mat,
  peaks,
  genome,
  id = NULL,
  description = NULL,
  metadata = NULL,
  cell_to_metacell = NULL,
  mc_size_eps_q = 0.1,
  path = "",
  tad_based = TRUE
)
}
\arguments{
\item{mat}{a numeric matrix where rows are peaks and columns are metacells. Can be a sparse matrix.}

\item{peaks}{misha intervals set. Can contain a field named 'peak_name' with a unique name per peak. Both the names and intervals should be unique (a peak cannot appear more than once).}

\item{genome}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{id}{an identifier for the object, e.g. "pbmc".}

\item{description}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k),
projection was done using RNA metacells"}

\item{metadata}{data frame with a column called 'metacell' and additional metacell annotations, or the name of a delimited file which contains such annotations.}

\item{cell_to_metacell}{a data frame mapping 'cell_id' to 'metacell' (optional). See \code{project_atac_on_mc}}

\item{mc_size_eps_q}{(optional) quantile of MC size (in UMIs) to scale the number of UMIs per metacell. \eqn{egc_ij} would then be
the fraction of peak i in metacell j multiplied by the \code{mc_size_eps_q} quantile of metacell sizes.}

\item{path}{path from which the object was loaded (optional)}

\item{tad_based}{whether the peak names are TAD-based}
}
\description{
McATACPeaks is a shallow object holding ATAC data over metacells.
Minimally it should include a count matrix of peaks over metacells, and \code{PeakIntervals} which hold the coordinates
of the peaks.
}
