% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\name{initialize,ScATAC-method}
\alias{initialize,ScATAC-method}
\title{Construct a new ScATAC object}
\usage{
\S4method{initialize}{ScATAC}(.Object, mat, peaks, genome, metadata = NULL)
}
\arguments{
\item{mat}{a numeric matrix where rows are peaks and columns are cells. Can be a sparse matrix.}

\item{peaks}{misha intervals set. Can contain a field named 'peak_name' with a unique name per peak. Both the names and intervals should be unique (a peak cannot appear more than once).}

\item{genome}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{metadata}{data frame with a column called 'cell_id' and additional per-cell annotations, or the name of a delimited file which contains such annotations.}
}
\description{
ScATAC is a shallow object holding ATAC data over cells.
Minimally it should include a count matrix of peaks over cells, and \code{PeakIntervals} which hold the coordinates
of the peaks.
}
