% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACPeaks.R
\name{initialize,ScPeaks-method}
\alias{initialize,ScPeaks-method}
\title{Construct a new ScPeaks object}
\usage{
\S4method{initialize}{ScPeaks}(
  .Object,
  mat,
  peaks,
  genome,
  id = NULL,
  description = NULL,
  metadata = NULL,
  path = "",
  tad_based = TRUE
)
}
\arguments{
\item{mat}{a numeric matrix where rows are peaks and columns are cells. Can be a sparse matrix.}

\item{peaks}{misha intervals set. Can contain a field named 'peak_name' with a unique name per peak. Both the names and intervals should be unique (a peak cannot appear more than once).}

\item{genome}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{id}{an identifier for the object, e.g. "pbmc".}

\item{description}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"}

\item{metadata}{data frame with a column called 'cell_id' and additional per-cell annotations, or the name of a delimited file which contains such annotations.}

\item{path}{path from which the object was loaded (optional)}

\item{tad_based}{wether the peak names are TAD-based}
}
\description{
ScPeaks is a shallow object holding ATAC data over cells.
Minimally it should include a count matrix of peaks over cells, and \code{PeakIntervals} which hold the coordinates
of the peaks.
}
