% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{mc_hclust_rna}
\alias{mc_hclust_rna}
\title{Hierarchical clustering of metacells using RNA markers}
\usage{
mc_hclust_rna(
  atac_mc,
  markers = NULL,
  force_cell_type = TRUE,
  rm_zeros = TRUE,
  epsilon = 0.00001,
  ...
)
}
\arguments{
\item{atac_mc}{a McPeaks/McTracks object with RNA expression (using \code{add_mc_rna})}

\item{markers}{a list of marker genes. If NULL - the function uses \code{get_rna_markers} with default parameters which can be overridden
using the ellipsis \code{...}.}

\item{force_cell_type}{do not split cell types when ordering the metacells. Default: TRUE}

\item{rm_zeros}{remove genes with no RNA expression in any metacell. Default: TRUE}

\item{epsilon}{regularization factor added to the log normalized expression}

\item{...}{
  Arguments passed on to \code{\link[=get_rna_markers]{get_rna_markers}}
  \describe{
    \item{\code{n_genes}}{maximal number of genes to return. Default: 100}
    \item{\code{minimal_max_log_fraction}}{take only genes with at least one value
(in log fraction units - normalized egc) above this threshold}
    \item{\code{minimal_relative_log_fraction}}{take only genes with at least one value with relative
log fraction (mc_fp) above this this value}
    \item{\code{fold_change_reg}}{regularization factor for the fold change calculation (fold_change would be changed to
\code{fold_change = fold_change + fold_change_reg})}
    \item{\code{genes}}{list of genes to match. Default (NULL): all genes}
  }}
}
\value{
an 'hclust' object with the hierarchical clustering of the metacells. Note that the order of the metacells within the 'hclust' object
is not necessarily the same as the order of the metacells in the input metacells object. Use the 'labels' slot in order to infer it (see examples).
}
\description{
create an 'hclust' object using the RNA marker matrix. When \code{force_cell_type} is TRUE and \code{atac_mc@metadata} has a field named "cell_type" the clustering is done
}
\examples{
\dontrun{
mat <- get_rna_marker_matrix(atac_mc)
hc <- mc_hclust_rna(atac_mc)
mat <- mat[, hc$labels[hc$order], drop = FALSE]
gene_ord <- order(apply(mat, 1, which.max))
mat <- mat[gene_ord, , drop = FALSE]
plot_rna_markers_mat(mat, atac_mc@metadata, atac_mc@metadata, col_names = F)
}

}
