% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{mc_order_by_rna}
\alias{mc_order_by_rna}
\title{Order metacells by RNA marker clustering}
\usage{
mc_order_by_rna(
  atac_mc,
  markers = NULL,
  force_cell_type = TRUE,
  rm_zeros = TRUE,
  epsilon = 0.00001,
  ...
)
}
\arguments{
\item{atac_mc}{a McPeaks/McTracks object with RNA expression (using \code{add_mc_rna})}

\item{markers}{a list of marker genes. If NULL - the function uses \code{get_rna_markers} with default parameters which can be overridden
using the ellipsis \code{...}.}

\item{force_cell_type}{do not split cell types when ordering the metacells. Default: TRUE}

\item{rm_zeros}{remove genes with no RNA expression in any metacell. Default: TRUE}

\item{epsilon}{regularization factor added to the log normalized expression}

\item{...}{
  Arguments passed on to \code{\link[=get_rna_marker_matrix]{get_rna_marker_matrix}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
order the metacells using the RNA marker clustering. See \code{get_rna_marker_matrix} for more information.
When \code{force_cell_type} is TRUE and \code{atac_mc@metadata} has a field named "cell_type", the columns are only ordered only within each
cell type.
}
