% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\name{McATAC}
\alias{McATAC}
\title{Construct a new McATAC object}
\usage{
McATAC(mat, peaks, metadata = NULL)
}
\arguments{
\item{mat}{a numeric matrix where rows are peaks and columns are metacells. Can be a sparse matrix.}

\item{peaks}{misha intervals set. Can contain a field named 'peak_name' with a unique name per peak. Both the names and intervals should be unique (a peak cannot appear more than once).}

\item{metadata}{data frame with a column called 'metacell' and additional metacell annotations.}
}
\description{
McATAC is a shallow object holding ATAC data over metacells.
Minimally it should include a count matrix of peaks over metacells, and \code{PeakIntervals} which hold the coordinates
of the peaks.
}
