% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{mcc_extract_to_df}
\alias{mcc_extract_to_df}
\title{Extract McCounts to a tidy data frame}
\usage{
mcc_extract_to_df(mc_counts, metacells = NULL)
}
\arguments{
\item{mc_counts}{A McCounts object}

\item{metacells}{metacells to extract.}
}
\value{
an intervals set (tibble) with additional "metacell" and "value" columns
}
\description{
Extract McCounts to a tidy data frame
}
\examples{
\dontrun{
df <- mcc_extract_to_df(mc_counts, c("1", "2"))
}

}
