% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{mcc_marginal}
\alias{mcc_marginal}
\title{Return the total coverage of each non-zero coordinate in an McCounts object}
\usage{
mcc_marginal(mc_counts, metacells = NULL)
}
\arguments{
\item{mc_counts}{a McCounts object}

\item{metacells}{names of metacells to include. Default: all metacells.}
}
\value{
an intervals set with an additional column called "cov" with the total coverage for each coordinate
}
\description{
Return the total coverage of each non-zero coordinate in an McCounts object
}
\examples{
\dontrun{
mc_counts <- mcc_read("pbmc_reads_mc")
mc_mars <- mcc_marginal(mc_counts)
}

}
