% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{mcc_normalize_metacells}
\alias{mcc_normalize_metacells}
\title{Normalize each metacell in a McCounts object by its total counts}
\usage{
mcc_normalize_metacells(mc_counts, metacells = NULL)
}
\arguments{
\item{mc_counts}{a McCounts object}

\item{metacells}{metacells to normalize. Default: all metacells.}
}
\value{
an McCounts object with the metacells normalized
}
\description{
Normalize each metacell in a McCounts object by its total counts
}
\examples{
\dontrun{
mc_counts <- mcc_read("pbmc_reads_mc")
mc_counts_norm <- mcc_normalize(mc_counts)
}

}
