% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{mcc_to_mcatac}
\alias{mcc_to_mcatac}
\title{Create an McATAC object from an McCounts object}
\usage{
mcc_to_mcatac(
  mc_counts,
  peaks,
  metacells = NULL,
  metadata = NULL,
  mc_size_eps_q = 0.1
)
}
\arguments{
\item{mc_counts}{a McCounts object}

\item{peaks}{a data frame with the peak intervals (chrom, start, end) and a column called "peak_name"}

\item{metacells}{names of metacells to include. Default: all metacells.}

\item{metadata}{(optional) per-metacell metadata. A data frame with a column called 'metacell' and additional metacell annotations.}

\item{mc_size_eps_q}{(optional) quantile of MC size (in UMIs) to scale the number of UMIs per metacell. \eqn{egc_ij} would then be
the fraction of peak i in metacell j multiplied by the \code{mc_size_eps_q} quantile of metacell sizes.}
}
\value{
a McATAC object
}
\description{
given an McCounts object and peaks, summarise the counts over the peaks and return a McATAC object
}
\examples{
\dontrun{
atac_sc <- import_from_10x("pbmc_data", genome = "hg38")
mc_counts <- mcc_read("pbmc_reads_mc")
atac_mc <- mcc_to_mcatac(mc_counts, atac_sc@peaks)
}

}
