% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{mcc_to_tracks}
\alias{mcc_to_tracks}
\title{Create a misha track for each metacell in a McCounts object}
\usage{
mcc_to_tracks(
  mc_counts,
  track_prefix,
  metacells = NULL,
  overwrite = FALSE,
  normalize = TRUE
)
}
\arguments{
\item{mc_counts}{A McCounts object}

\item{track_prefix}{The prefix of the tracks to create. Track names will be of the form "{track_prefix}.mc{metacell}"}

\item{metacells}{metacells for which to create tracks. If NULL, all metacells will be used.}

\item{overwrite}{Whether to overwrite existing tracks.}

\item{normalize}{Normalize each metacell by the sum of its counts.}
}
\value{
None.
}
\description{
This would create a sparse track of the form "{track_prefix}.mc{metacell}" for each metacell in the McCounts object. By default, the counts are normalized to the total counts of each metacell, set \code{normalize = FALSE}
to disable this.
}
\examples{
\dontrun{
mcc_to_tracks(mc_counts, "pbmc_mc")
}

}
