% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACTracks.R
\name{mct_create}
\alias{mct_create}
\title{Create an McTracks object}
\usage{
mct_create(
  genome,
  tracks = NULL,
  track_prefix = NULL,
  metacells = NULL,
  marginal_track = NULL,
  id = NULL,
  description = NULL,
  path = NULL,
  metadata = NULL
)
}
\arguments{
\item{genome}{genome assembly of the tracks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{tracks}{a vector of tracks to include in the object.}

\item{track_prefix}{prefix for tracks (instead of explicit track names at the \code{tracks} parameter). Ignored if \code{tracks} is not NULL. If
a track named "{track_prefix}.marginal" exists, it will be used as the marginal track.}

\item{metacells}{a vector of metacells to include in the object (optional). If not provided, the metacells are inferred from the track names using the pattern "{track_prefix}.mc\\d*$". In such a case, the tracks are ordered according to the metacell number.}

\item{marginal_track}{a track to use as the marginal track (optional). If not provided, and \code{track_prefix} was given, the track named "{track_prefix}.marginal" will be used if it exists.}

\item{id}{an identifier for the object, e.g. "pbmc".}

\item{description}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"}

\item{path}{path to a directory containing the raw data. Usually inherited from the counts object.}

\item{metadata}{data frame with a column called 'metacell' and additional metacell annotations, or the name of a delimited file which contains such annotations.}
}
\value{
a McTracks object.
}
\description{
Create an McTracks object either from a list of tracks (\code{tracks} parameter) or from a track prefix (\code{track_prefix} parameter). The tracks are assumed to be in the form of ".+_mc{metacell_number}", if this is not the case, the \code{metacells} parameter can be used to specify the metacell number for each track.
}
\examples{
\dontrun{
mct_create(track_prefix = "pbmc_mc", genome = "hg38", id = "pbmc", description = "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)")
}

}
