% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACTracks-cache.R
\name{mct_get_mat}
\alias{mct_get_mat}
\title{Get an ATAC matrix from a McTracks object}
\usage{
mct_get_mat(mct, intervals, downsample = FALSE, force = FALSE)
}
\arguments{
\item{mct}{a McTrack object}

\item{downsample}{return a downsampled matrix. Downsampling is done by
TODO}

\item{force}{force the computation of the matrix. If FALSE, the matrix is
retrieved from the cache if it exists.}

\item{intervlas}{an intervals set}
}
\value{
a matrix where rows are genomic coordinates in the resolution
of \code{mct@resolution} and columns are the metacells in \code{mct@metacells}
}
\description{
Get an ATAC matrix from a McTracks object
}
