% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACTracks-cache.R
\name{mct_ls_cached_regions}
\alias{mct_ls_cached_regions}
\title{List the cached regions of an McTracks object}
\usage{
mct_ls_cached_regions(mct)
}
\arguments{
\item{mct}{a McTracks object}
}
\value{
an intervals set with additional "downsample" and "downsample_n" fields, representing the intervals matrices that are currently cached.
}
\description{
List the cached regions of an McTracks object
}
\examples{
\dontrun{
mct_ls_cached_regions(mct)
}

}
