% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACTracks.R
\name{mct_to_mcatac}
\alias{mct_to_mcatac}
\title{Create an McPeaks object from McTracks object}
\usage{
mct_to_mcatac(
  mct,
  peaks,
  metacells = NULL,
  metadata = NULL,
  mc_size_eps_q = 0.1,
  parallel = getOption("mcatac.parallel")
)
}
\arguments{
\item{mct}{an McTracks object}

\item{peaks}{a data frame with the peak intervals (chrom, start, end) and a column called "peak_name"}

\item{metacells}{names of metacells to include. Default: all metacells.}

\item{metadata}{(optional) per-metacell metadata. A data frame with a column called 'metacell' and additional metacell annotations.}

\item{mc_size_eps_q}{(optional) quantile of MC size (in UMIs) to scale the number of UMIs per metacell. \eqn{egc_ij} would then be
the fraction of peak i in metacell j multiplied by the \code{mc_size_eps_q} quantile of metacell sizes.}
}
\value{
a McPeaks object
}
\description{
Create an McPeaks object from an McTracks object by extracting the atac data from the tracks at the peaks.
}
\examples{
\dontrun{
atac_mc <- mct_to_mcatac(mct, peaks)
}

}
