% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr-bias.R
\name{normalize_const}
\alias{normalize_const}
\title{Normalize McPeaks data using constitutive peaks}
\usage{
normalize_const(mcatac, norm_type, norm_quant = 1)
}
\arguments{
\item{mcatac}{The McPeaks object containing the data to be normalized. The object should contain a field named {.field const} in {.field peaks}. Use \code{add_const_peaks} to add this field.}

\item{norm_type}{The cell type to use as the reference for normalization}

\item{norm_quant}{The quantile to which the normalized data should be scaled (default is 1).}
}
\value{
The McPeaks object with normalized egc values.
}
\description{
The function normalizes the McPeaks data using the values of the mean of constitutive peaks in a specified cell type. The normalized data is then scaled to a specified quantile and capped at 1.
}
\examples{
\dontrun{
mcatac <- import_from_matrix(mat, peaks, genome = "mm10", class = "McPeaks")
mcatac <- normalize_egc(mcatac, "gastrulation.marginal")
mcatac <- add_const_peaks(mcatac, -16)
mcatac <- normalize_const(mcatac, norm_type = "Epiblast")
}

}
