% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr-bias.R
\name{normalize_egc}
\alias{normalize_egc}
\title{Normalize McPeaks data using marginal coverage}
\usage{
normalize_egc(
  mcatac,
  marginal_track,
  window_size = 10000,
  epsilon = 0.00001,
  minimal_quantile = 0.1
)
}
\arguments{
\item{mcatac}{The McPeaks object containing the data to be normalized.}

\item{marginal_track}{The marginal track}

\item{epsilon}{The epsilon value added to the egc matrix (default is 1e-5).}

\item{widnow_size}{The size of the windows to use around each peak for normalization.}
}
\value{
The McPeaks object with normalized egc values.
}
\description{
The function calculates the (punctured) marginal coverage in \code{window_size} windows around each peak,
and then normalizes the data by dividing the original data by the punctured marginal coverage, while setting
the minimum to \code{minimal_quantile} of the punctured marginal coverage.
}
\examples{
\dontrun{
mcatac <- import_from_matrix(mat, peaks, genome = "hg38", class = "McPeaks")
mcatac <- normalize_egc(mcatac, "pbmc_mc.marginal")
}

}
