% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr-bias.R
\name{normalize_to_prob}
\alias{normalize_to_prob}
\title{Normalize the mcATAC object to probability values}
\usage{
normalize_to_prob(mcatac, prob1_thresh = NULL, const_quantile = 0.8)
}
\arguments{
\item{mcatac}{The mcPeaks object to be normalized.}

\item{prob1_thresh}{The threshold value for setting values above it to the threshold value. If not provided, the function will calculate the threshold using the "const" field in the peaks.}

\item{const_quantile}{The quantile value used to calculate the threshold when "prob1_thresh" is not provided.}
}
\value{
The normalized mcATAC object.
}
\description{
This function normalizes the mcATAC object to probability values by setting values above a threshold to the threshold value,
and then scaling the values to the range \link{0, 1}.
}
\examples{
\dontrun{
mcatac <- import_from_matrix(mat, peaks, genome = "mm10", class = "McPeaks")
mcatac <- normalize_egc(mcatac, "gastrulation.marginal")
mcatac <- add_const_peaks(mcatac, -16)
mcatac <- normalize_to_prob(mcatac)
}

}
