% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_tracks}
\alias{order_tracks}
\title{Order metacell tracks by metacell number}
\usage{
order_tracks(tracks, track_name_prefix = NULL)
}
\arguments{
\item{tracks}{the tracks to order}

\item{track_name_prefix}{prefix string to remove when extracting metacell numbers}
}
\value{
tracks ordered by metacell number
}
\description{
Order metacell tracks by metacell number
}
\examples{
tracks_ordered <- order_tracks(tracks)
}
