% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakIntervals.R
\name{peak_names}
\alias{peak_names}
\title{Return the names of PeakIntervals}
\usage{
peak_names(peaks)
}
\arguments{
\item{peaks}{a PeakIntervals object or misha intervals}
}
\value{
the field called 'peak_name' if exists in \code{peaks}, or the coordinates seperated by an underscore.
}
\description{
Return the names of PeakIntervals
}
