% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_atac_atac_cor}
\alias{plot_atac_atac_cor}
\title{Plot a correlation matrix of ATAC metacells}
\usage{
plot_atac_atac_cor(mc_atac)
}
\arguments{
\item{mc_atac}{McATAC object}
}
\description{
Plot a correlation matrix of ATAC metacells
}
