% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_atac_rna}
\alias{plot_atac_rna}
\title{Plot a scatter of gene expression vs an atac profile}
\usage{
plot_atac_rna(mc_atac, gene, peak = NULL)
}
\arguments{
\item{mc_atac}{a McATAC object}

\item{gene}{name of the gene top plot}

\item{peak}{name of the peak to plot. If NULL - the promoter of \code{gene} would be shown}
}
\value{
a ggplot object with the scatter plot
}
\description{
Plot a scatter of gene expression vs an atac profile
}
