% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_atac_rna_cor}
\alias{plot_atac_rna_cor}
\title{Plot a cross-correlation matrix between RNA metacells and ATAC metacell scores}
\usage{
plot_atac_rna_cor(mc_atac, rna_mat)
}
\arguments{
\item{mc_atac}{McATAC object}

\item{rna_mat}{an RNA metacell count matrix, where metacells are in columns and genes are in rows}

\item{gene_field}{name of a field in \code{mc_atac$peaks} which contains the gene names. If NULL - the peaks would be
transformed to promoter peaks and the gene names would be taken from the promoter gene names.}
}
\description{
use peak gene annotations to match between RNA metacells and ATAC metacells and plot
a cross-correlation matrix.
}
