% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-markers.R
\name{plot_atac_rna_markers}
\alias{plot_atac_rna_markers}
\title{Plot RNA and ATAC marker matrix side by side}
\usage{
plot_atac_rna_markers(
  atac_mc,
  n_genes = 100,
  force_cell_type = TRUE,
  plot_legend = TRUE,
  row_names = FALSE,
  ...
)
}
\arguments{
\item{atac_mc}{a McATAC object with RNA expression (using \code{add_mc_rna})}

\item{n_genes}{maximal number of genes to return. Default: 1000}

\item{force_cell_type}{do not split cell types when ordering the metacells. Default: TRUE}

\item{plot_legend}{TRUE to plot a legend of the cell types (default: TRUE)}

\item{row_names}{show the row names of the heatmaps. Note that the space each plot takes might become uneven due to
genes / peaks with longer names.}

\item{...}{
  Arguments passed on to \code{\link[=plot_rna_markers_mat]{plot_rna_markers_mat}}
  \describe{
    \item{\code{mat}}{a heatmap of marker genes fold change over metacells (e.g. output of \code{get_rna_marker_matrix})}
    \item{\code{metacell_types}}{a data frame with a field called "metacell" and a field called "cell_type" (optional)}
    \item{\code{cell_type_colors}}{a data frame with a field called "cell_type" and a field called "color" (optional)}
    \item{\code{low_color}}{a color for low fold change}
    \item{\code{high_color}}{a color for high fold change}
    \item{\code{mid_color}}{a color for mid fold change}
    \item{\code{midpoint}}{midpoint for the color scale (default: 0)}
    \item{\code{min_lfp}}{minimum log2 fold change (default: -3). Value below this will be set to \code{min_lfp}.}
    \item{\code{max_lfp}}{maximum log2 fold change (default: 3). Value above this will be set to \code{max_lfp}.}
    \item{\code{top_cell_type_bar}}{add a cell type annotation also at the top of the heatmap (default: TRUE)}
    \item{\code{gene_colors}}{a named list with a color for each marker gene (optional)}
    \item{\code{col_names}}{show column names (metacells)}
    \item{\code{interleave}}{show the gene names (rows) interleaved}
    \item{\code{annotation_title}}{title for cell type annotation. Default: "Cell type"}
    \item{\code{top_annotation_title}}{title for top cell type annotation. Default: "Cell type"}
    \item{\code{title}}{title for the plot}
    \item{\code{plot_right}}{Whether to put the Y-axis labels at the right side of the heatmap (as well as at the left hand side)}
    \item{\code{col_names_orient}}{Orientation of the X-axis label}
  }}
}
\value{
a \code{ggplot} object with the heatmap of RNA markers (right) and the heatmap of ATAC peaks (left).
The ATAC matrix is ordered by the RNA matrix.

a ggplot object with the heatmaps
}
\description{
plot an rna marker heatmap (right) together with an heatmap of fold change of ATAC peaks that
are the most correlated to each marker gene (left, using \code{get_genes_atac_fp}).
}
\examples{
\dontrun{
plot_atac_rna_markers(atac_mc)
plot_atac_rna_markers(atac_mc, force_cell_type = FALSE)
plot_atac_rna_markers(atac_mc, n_genes = 200)
}

}
