% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{plot_motif_logo}
\alias{plot_motif_logo}
\title{Plot LOGO of a motif}
\usage{
plot_motif_logo(dataset, motif)
}
\arguments{
\item{dataset}{name of the dataset from which the motif came from, e.g. "homer", "jolma"}

\item{motif}{name of the motif to plot, e.g. "GATA"}
}
\value{
a ggplot2 object
}
\description{
Plot LOGO of a motif
}
\examples{
\dontrun{
gset_genome("mm10")
plot_motif_logo("homer", "GATA") #'
}

}
