% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.r
\name{plot_peak_coverage_density}
\alias{plot_peak_coverage_density}
\title{Plot the coverage density of the peaks}
\usage{
plot_peak_coverage_density(atac, scale = 100, pointsize = 1.5)
}
\arguments{
\item{atac}{an ScATAC or McATAC object}

\item{scale}{for the 'cov_density' field, calculate the number of UMIs per \code{scale}
bp (default: 100)}

\item{pointsize}{Radius of rasterized point. Use \code{0} for single pixels (fastest).}
}
\description{
For each peak, plot a scatter of the peak length vs the number
of UMIs per \code{scale} bp.
}
\examples{
\dontrun{
plot_peak_coverage_density(atac)
}

}
