% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{plot_peak_coverage_distribution}
\alias{plot_peak_coverage_distribution}
\title{Plot the coverage distribution of the peaks}
\usage{
plot_peak_coverage_distribution(atac)
}
\arguments{
\item{atac}{an ScATAC or McATAC object}
}
\value{
a ggplot object with the distribution of peak coverage
}
\description{
Plot the coverage distribution of the peaks
}
\examples{
\dontrun{
plot_peak_coverage_distribution(scatac)
}
}
