% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{plot_peak_length_distribution}
\alias{plot_peak_length_distribution}
\title{Plot the distribution of the peak length}
\usage{
plot_peak_length_distribution(atac)
}
\arguments{
\item{atac}{an ScATAC or McATAC object}
}
\value{
a ggplot object with the distribution of peak length
}
\description{
Plot the distribution of the peak length
}
\examples{
\dontrun{
plot_peak_length_distribution(atac_sc)
}

}
