% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{plot_peak_max_cov_distribution}
\alias{plot_peak_max_cov_distribution}
\title{Plot the maximal per-cell coverage distribution of the peaks}
\usage{
plot_peak_max_cov_distribution(atac)
}
\arguments{
\item{atac}{an ScATACPeaks or McATACPeaks object}
}
\value{
a ggplot object with the distribution of maximal per-cell coverage for each peak
}
\description{
Plot the maximal per-cell coverage distribution of the peaks
}
\examples{
\dontrun{
plot_peak_max_cov_distribution(atac_sc)
}

}
