% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-markers.R
\name{plot_rna_markers}
\alias{plot_rna_markers}
\title{Plot a heatmap of marker genes fold change over metacells given an McATAC object}
\usage{
plot_rna_markers(atac_mc, n_genes = 100, force_cell_type = TRUE, ...)
}
\arguments{
\item{atac_mc}{a McATAC object with RNA expression (using \code{add_mc_rna})}

\item{n_genes}{maximal number of genes to return. Default: 1000}

\item{force_cell_type}{do not split cell types when ordering the metacells. Default: TRUE}

\item{...}{
  Arguments passed on to \code{\link[=plot_rna_markers_mat]{plot_rna_markers_mat}}
  \describe{
    \item{\code{mat}}{a heatmap of marker genes fold change over metacells (e.g. output of \code{get_rna_marker_matrix})}
    \item{\code{metacell_types}}{a data frame with a field called "metacell" and a field called "cell_type" (optional)}
    \item{\code{cell_type_colors}}{a data frame with a field called "cell_type" and a field called "color" (optional)}
    \item{\code{low_color}}{a color for low fold change}
    \item{\code{high_color}}{a color for high fold change}
    \item{\code{mid_color}}{a color for mid fold change}
    \item{\code{midpoint}}{midpoint for the color scale (default: 0)}
    \item{\code{min_lfp}}{minimum log2 fold change (default: -3). Value below this will be set to \code{min_lfp}.}
    \item{\code{max_lfp}}{maximum log2 fold change (default: 3). Value above this will be set to \code{max_lfp}.}
    \item{\code{plot_legend}}{TRUE to plot a legend of the cell types (default: TRUE)}
    \item{\code{top_cell_type_bar}}{add a cell type annotation also at the top of the heatmap (default: TRUE)}
    \item{\code{gene_colors}}{a named list with a color for each marker gene (optional)}
    \item{\code{col_names}}{show column names (metacells)}
    \item{\code{interleave}}{show the gene names (rows) interleaved}
    \item{\code{title}}{title for the plot}
  }}
}
\description{
This is a thin wrapper around plot_rna_markers_mat which computes the marker matrix and uses the correct
metadata fields in the McATAC object.
}
\examples{
\dontrun{
plot_rna_markers(atac_mc)
}
}
