% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-markers.R
\name{plot_rna_markers_mat}
\alias{plot_rna_markers_mat}
\title{Plot a heatmap of marker genes fold change over metacells}
\usage{
plot_rna_markers_mat(
  mat,
  metacell_types = NULL,
  cell_type_colors = NULL,
  low_color = "blue",
  high_color = "red",
  mid_color = "white",
  midpoint = 0,
  min_lfp = -3,
  max_lfp = 3,
  plot_legend = TRUE,
  top_cell_type_bar = TRUE,
  gene_colors = NULL,
  col_names = FALSE,
  row_names = TRUE,
  col_names_orient = "slanted",
  plot_right = TRUE,
  interleave = TRUE,
  vertial_gridlines = FALSE,
  annotation_title = "Cell type",
  top_annotation_title = "Cell type",
  title = NULL
)
}
\arguments{
\item{mat}{a heatmap of marker genes fold change over metacells (e.g. output of \code{get_rna_marker_matrix})}

\item{metacell_types}{a data frame with a field called "metacell" and a field called "cell_type" (optional)}

\item{cell_type_colors}{a data frame with a field called "cell_type" and a field called "color" (optional)}

\item{low_color}{a color for low fold change}

\item{high_color}{a color for high fold change}

\item{mid_color}{a color for mid fold change}

\item{midpoint}{midpoint for the color scale (default: 0)}

\item{min_lfp}{minimum log2 fold change (default: -3). Value below this will be set to \code{min_lfp}.}

\item{max_lfp}{maximum log2 fold change (default: 3). Value above this will be set to \code{max_lfp}.}

\item{plot_legend}{TRUE to plot a legend of the cell types (default: TRUE)}

\item{top_cell_type_bar}{add a cell type annotation also at the top of the heatmap (default: TRUE)}

\item{gene_colors}{a named list with a color for each marker gene (optional)}

\item{col_names}{show column names (metacells)}

\item{row_names}{The names that will be used as the Y-axis labels.
A NULL (the defualt) indicates that the original matrix rownames are used.
A FALSE indicates that no Y-axis labels should be displayed.}

\item{col_names_orient}{Orientation of the X-axis label}

\item{plot_right}{Whether to put the Y-axis labels at the right side of the heatmap (as well as at the left hand side)}

\item{interleave}{show the gene names (rows) interleaved}

\item{annotation_title}{title for cell type annotation. Default: "Cell type"}

\item{top_annotation_title}{title for top cell type annotation. Default: "Cell type"}

\item{title}{title for the plot}

\item{vertical_gridlines}{show vertical gridlines}
}
\value{
a ggplot object with the heatmap (and legend if \code{plot_legend} is TRUE)
}
\description{
Plot a heatmap of marker genes fold change over metacells
}
\examples{
\dontrun{
m_rna <- get_rna_marker_matrix(atac_mc)
plot_rna_markers_mat(m_rna, atac_mc@metadata, atac_mc@metadata, col_names = F)
}
}
