% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tracks_at_locus}
\alias{plot_tracks_at_locus}
\title{Plot metacell tracks around locus}
\usage{
plot_tracks_at_locus(
  tracks = NULL,
  gene = NULL,
  intervals = NULL,
  mc_rna = NULL,
  gene_feature = "exon",
  track_regex = NULL,
  iterator = 100,
  extend = 0,
  order_rows = FALSE,
  row_order = NULL,
  gene_annot = TRUE,
  annotation_row = NULL,
  annotation_col = NULL,
  annotation_colors = NULL,
  silent = FALSE,
  rna_legc_eps = 1e-05,
  colors = colorRampPalette(c("white", "darkblue", "red"))(100),
  ...
)
}
\arguments{
\item{tracks}{(optional) all tracks to plot}

\item{gene}{(optional) which gene to plot around}

\item{intervals}{(optional) what genomic interval to plot}

\item{mc_rna}{(optional) RNA metacell object to extract gene expression data from}

\item{track_regex}{(optional) regular expression for matching tracks to plot}

\item{iterator}{(optional) misha iterator}

\item{extend}{(optional) how much to extend \code{intervals} by on each side}

\item{gene_annot}{(optional) whether to add gene annotations}

\item{annotation_row}{pheatmap-format annotation}

\item{annotation_col}{pheatmap-format annotation}

\item{annotation_colors}{pheatmap-format annotation}

\item{silent}{(optional) whether to print generated plot}

\item{rna_legc_eps}{(optional) what regularization value to add to mc_rna@e_gc when calculating log}

\item{colors}{(optional) colorRampPalette vector of colors for scaling colors in heatmap}

\item{...}{
  Arguments passed on to \code{\link[=save_pheatmap]{save_pheatmap}}
  \describe{
    \item{\code{dev}}{name of the device to save the pheatmap. e.g. "png" or "pdf"}
    \item{\code{filename}}{the output file path.
    The page number is substituted if a C integer format is included in
    the character string, as in the default.  (The result must be less
    than \code{PATH_MAX} characters long, and may be truncated if not.
    See \code{\link[grDevices]{postscript}} for further details.)  Tilde expansion
    is performed where supported by the platform.  An input with
    a marked encoding is converted to the native encoding on an error is
    given.}
    \item{\code{width}}{the width of the device.}
    \item{\code{height}}{the height of the device.}
    \item{\code{res}}{The nominal resolution in ppi which will be recorded in the
    bitmap file, if a positive integer.  Also used for \code{units}
    other than the default, and to convert points to pixels.}
  }}

\item{atac}{(optional) ScATAC, McATAC or PeakIntervals object from which to extract peaks in locus}
}
\value{
a ComplexHeatmap figure
}
\description{
Plot metacell tracks around locus
}
\examples{
\dontrun{
    library(metacell)
    scdb_init("scdb")
    mc_rna = scdb_mc('rna_w_color_key')
    mcmd = readr::read_csv('./data/mcmd.csv')
    color_key = unique(mcmd[,c('st', 'color')])
    colnames(color_key) = c('cell_type', 'color')
    col_annot = data.frame(cell_type = mcmd$st)
    rownames(col_annot) = mcmd$mc
    ann_colors = list(cell_type = setNames(color_key$color, color_key$cell_type))
    plot_tracks_at_locus(tracks = pbmc_tracks, extend = 5e+4,
                    gene = "ATF3", 
                    mc_rna = mc_rna, 
                    gene_annot = T, 
                    order_rows = T,
                    annotation_row = col_annot, 
                    annotation_colors = ann_colors)
}
}
