% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{prepare_clusters}
\alias{prepare_clusters}
\title{Generate a ucsc genome browser file for each metacell cluster}
\usage{
prepare_clusters(mc_atac, clust_vec = NULL, normalization = "none")
}
\arguments{
\item{mc_atac}{McATAC object}

\item{clust_vec}{(optional) a vector of length #metacells representing an annotation/clustering
(can be output of \code{gen_atac_mc_clust})}

\item{normalization}{(optional) normalization method, either 'none', 'lfcom' (log2 fold-change over median), 'zs' (Z-scores)}

\item{track_prefix}{prefix for generated misha tracks.}

\item{output_dir}{(optional) name of the directory to write files to}
}
\description{
Generate a ucsc genome browser file for each metacell cluster
}
