% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_on_mc.R
\name{project_atac_on_mc}
\alias{project_atac_on_mc}
\alias{project_atac_on_mc_from_metacell1}
\alias{project_atac_on_mc_from_h5ad}
\title{Given metacells (usually from RNA data), project ATAC counts to get a McATAC object}
\usage{
project_atac_on_mc(
  atac,
  mc = NULL,
  cell_to_metacell = NULL,
  MIN_INT_FRAC = 0.1
)

project_atac_on_mc_from_metacell1(atac, mc = NULL, cell_to_metacell = NULL)

project_atac_on_mc_from_h5ad(atac, h5ad_file)
}
\arguments{
\item{atac}{an ScATAC object}

\item{mc}{a metacell1 \code{mc} object}

\item{cell_to_metacell}{a data frame with a column named "cell_id" with cell id and}

\item{MIN_INT_FRAC}{(optional) minimal expected fraction of intersection of barcodes (cell names) in ScATAC and mc/cell_to_metacell
another column named "metacell" with the metacell the cell is part of.}

\item{h5ad_file}{name of an h5ad file which is the output of 'metacells' python package.}
}
\description{
Given cell to metacell association, summarise atac counts to generate a McATAC object.
\code{project_atac_on_mc_from_metacell1} projects given a metacell1 'mc' object, while \code{project_atac_on_mc_from_h5ad}
uses the output of the 'metacells' python package (metacell2).
}
