% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{project_counts_on_mc}
\alias{project_counts_on_mc}
\title{Given metacells (usually from RNA data), project ATAC counts to get a McCounts object}
\usage{
project_counts_on_mc(
  sc_counts,
  cell_to_metacell,
  num_cores = parallel::detectCores()
)
}
\arguments{
\item{sc_counts}{A ScCounts object}

\item{cell_to_metacell}{a data frame with a column named "cell_id" with cell id and another column named "metacell" with the metacell the cell is part of.}

\item{num_cores}{The number of cores to use.}
}
\value{
A McCounts object
}
\description{
Given metacells (usually from RNA data), project ATAC counts to get a McCounts object
}
\examples{
\dontrun{
data(cell_to_metacell_pbmc_example)
project_counts_on_mc(sc_counts, cell_to_metacell_pbmc_example)
}

}
