% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCounts.R
\name{read_mc_counts}
\alias{read_mc_counts}
\title{Read an McCounts object from a directory}
\usage{
read_mc_counts(path, id = NULL, description = NULL)
}
\arguments{
\item{path}{path to the directory containing the object (which was created by \code{write_sc_counts_from_bam})}

\item{id}{an identifier for the object (optional)}

\item{description}{description of the object (optional)}
}
\description{
Read an McCounts object from a directory
}
\examples{
\dontrun{
mc_counts <- read_sc_counts("pbmc_reads_mc")
}

}
