% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{rna_atac_cor_knn}
\alias{rna_atac_cor_knn}
\title{Match every gene with the k ATAC peaks most correlated to it}
\usage{
rna_atac_cor_knn(
  atac_mc,
  k = 1,
  genes = NULL,
  rm_zeros = TRUE,
  spearman = TRUE,
  pairwise.complete.obs = TRUE
)
}
\arguments{
\item{atac_mc}{a McPeaks/McTracks object with RNA expression (using \code{add_mc_rna})}

\item{k}{number of peaks to match for each gene. Default: 1}

\item{genes}{list of genes to match. Default (NULL): all genes}

\item{rm_zeros}{remove genes with no RNA expression in any metacell. Default: TRUE}

\item{spearman}{if 'TRUE' Spearman correlation is computed, otherwise
Pearson}

\item{pairwise.complete.obs}{see below}
}
\value{
a tibble with the following columns:
\itemize{
\item{gene: }{Gene name}
\item{peak: }{ATAC peak name}
\item{cor: }{Correlation between ATAC and RNA}
\item{rank: }{Rank of the correlation (for the gene)}
}
}
\description{
Match every gene with the k ATAC peaks most correlated to it
}
\examples{
\dontrun{
rna_atac_cor_knn(atac_mc, k = 1)
rna_atac_cor_knn(atac_mc, k = 1, genes = c("MESP1", "MESP2", "PF4"))
}

}
