% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCounts.R
\name{scc_change_names}
\alias{scc_change_names}
\title{Change the cell names of a ScCounts object}
\usage{
scc_change_names(sc_counts, new_names)
}
\arguments{
\item{sc_counts}{a ScCounts object}

\item{new_names}{new cell names}
}
\value{
a new ScCounts object with the new cell names
}
\description{
Change the cell names of a ScCounts object
}
\examples{
\dontrun{
sc_counts <- scc_read("pbmc_reads")
sc_counts <- scc_change_names(sc_counts, paste0(sc_counts@cell_names, "_rep1"))
}

}
