% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-counts.R
\name{scc_from_fragments}
\alias{scc_from_fragments}
\alias{write_sc_counts_from_fragments}
\title{Write single cell counts from a 10x fragments file}
\usage{
scc_from_fragments(
  fragments_file,
  out_dir,
  cell_names,
  cell_suffix = NULL,
  genome = NULL,
  bin_size = 50000000,
  overwrite = FALSE,
  id = "",
  description = "",
  verbose = FALSE,
  tabix_bin = "tabix",
  chromosomes = gintervals.all()$chrom
)

write_sc_counts_from_fragments(
  fragments_file,
  out_dir,
  cell_names,
  cell_suffix = NULL,
  genome = NULL,
  bin_size = 50000000,
  overwrite = FALSE,
  id = "",
  description = "",
  verbose = FALSE,
  tabix_bin = "tabix",
  chromosomes = gintervals.all()$chrom
)
}
\arguments{
\item{fragments_file}{path to fragments file. Note that in order to use tabix, the file must be compressed with gzip and have a ".gz" extension.}

\item{out_dir}{output directory.}

\item{cell_names}{a vector with the cell names or an ScPeaks object}

\item{cell_suffix}{a suffix to add to the cell names (optional). This is useful when you want to merge multiple batches of cells.
The suffix would be added in the following way: \code{paste0(cell_name, "-", cell_suffix)}.}

\item{bin_size}{Size of the genomic bins to use (in bp). Each chromsome will be chunked into bins with size which is
smaller than this value. Default is 50Mb.}

\item{overwrite}{overwrite existing directory (optional)}

\item{id}{an identifier for the object, e.g. "pbmc".}

\item{description}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"}

\item{verbose}{verbose output (optional)}

\item{tabix_bin}{path to the tabix binary (optional)}

\item{chromosomes}{process only specific chromosomes (optional).}
}
\value{
None
}
\description{
This function writes an ScCounts object from a "fragments.tsv.gz" which is an output of the 10x pipeline ("Barcoded and aligned fragment file").
}
\examples{
\dontrun{
scc_from_fragments("pbmc_data/fragments.tsv.gz", "pbmc_reads", cell_names = atac_sc)
scc_from_fragments("pbmc_data/fragments.tsv.gz", "pbmc_reads", cell_names = atac_sc, chromosomes = paste0("chr", c(1:22, "X", "Y")))
}

}
