% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{scc_project_on_mc}
\alias{scc_project_on_mc}
\title{Given metacells (usually from RNA data), project ATAC counts to get a McCounts object}
\usage{
scc_project_on_mc(sc_counts, cell_to_metacell, ignore_metacells = -1)
}
\arguments{
\item{sc_counts}{A ScCounts object}

\item{cell_to_metacell}{a data frame with a column named "cell_id" with cell id and another column named "metacell" with the metacell the cell is part of.}

\item{ignore_metacells}{a vector of metacells to ignore. Default: \link{-1} (the "outliers" metacell in the metacell2 python package).}
}
\value{
A McCounts object
}
\description{
Given cell to metacell association, summarise atac read counts to generate a McCounts object. This can
take a while - around 5 minutes using 24 cores on the PBMC dataset.
}
\examples{
\dontrun{
data(cell_to_metacell_pbmc_example)
scc_project_on_mc(sc_counts, cell_to_metacell_pbmc_example)
}

}
