% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCounts.R
\name{scc_to_peaks}
\alias{scc_to_peaks}
\title{Create an ScPeaks object from an ScCounts object}
\usage{
scc_to_peaks(
  sc_counts,
  peaks,
  cells = NULL,
  metadata = NULL,
  mc_size_eps_q = 0.1
)
}
\arguments{
\item{sc_counts}{a ScCounts object}

\item{peaks}{a data frame with the peak intervals (chrom, start, end) and a column called "peak_name"}

\item{cells}{names of cells to include. Default: all cells.}

\item{metadata}{(optional) per-metacell metadata. A data frame with a column called 'metacell' and additional metacell annotations.}

\item{mc_size_eps_q}{(optional) quantile of MC size (in UMIs) to scale the number of UMIs per metacell. \eqn{egc_ij} would then be
the fraction of peak i in metacell j multiplied by the \code{mc_size_eps_q} quantile of metacell sizes.}
}
\value{
a McPeaks object
}
\description{
given an ScCounts object and peaks, summarise the counts over the peaks and return a ScPeaks object
}
\examples{
\dontrun{
atac_sc <- import_from_10x("pbmc_data", genome = "hg38")
sc_counts <- scc_read("pbmc_reads")
atac_sc <- scc_to_peaks(sc_counts, atac_sc@peaks)
}

}
