% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{scc_to_scatac}
\alias{scc_to_scatac}
\title{Create a ScATAC object from an ScCounts object}
\usage{
scc_to_scatac(
  sc_counts,
  peaks,
  cells = NULL,
  metadata = NULL,
  mc_size_eps_q = 0.1,
  filter_intervs = NULL
)
}
\arguments{
\item{sc_counts}{a ScCounts object}

\item{peaks}{a data frame with the peak intervals (chrom, start, end) and a column called "peak_name"}

\item{cells}{name of the cells to include. Default: all cells.}

\item{metadata}{(optional) per-metacell metadata. A data frame with a column called 'metacell' and additional metacell annotations.}

\item{mc_size_eps_q}{(optional) quantile of MC size (in UMIs) to scale the number of UMIs per metacell. \eqn{egc_ij} would then be
the fraction of peak i in metacell j multiplied by the \code{mc_size_eps_q} quantile of metacell sizes.}

\item{filter_intervs}{intervals to filter out. The result would be the sum of the counts at \code{peaks} without the counts at \code{filter_intervs}. (optional)}
}
\value{
a ScATAC object
}
\description{
given an ScCounts object and peaks, summarise the counts over the peaks and return a ScATAC object
}
\examples{
\dontrun{
sc_counts <- sc_read("pbmc_reads")
sc_atac <- sc_to_scatac(sc_counts, atac_sc@peaks)
}

}
