% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCounts.R
\name{scc_write}
\alias{scc_write}
\title{Write a counts object to a file.}
\usage{
scc_write(object, out_dir, num_cores = parallel::detectCores())
}
\arguments{
\item{object}{a counts object (ScCounts or McCounts)}

\item{out_dir}{the output directory}

\item{num_cores}{the number of cores to use}
}
\description{
Write a counts object to a file.
}
\examples{
\dontrun{
scc_write(sc_counts, "sc_counts")
}

}
