% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCounts.R
\docType{class}
\name{ScCounts-class}
\alias{ScCounts-class}
\alias{ScCounts}
\title{ScCounts object}
\description{
ScCounts is a data structure that contains the counts of reads for a set of cells in a set of genomic bins.
}
\section{Slots}{

\describe{
\item{\code{id}}{an identifier for the object, e.g. "pbmc".}

\item{\code{description}}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"}

\item{\code{genome}}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{\code{data}}{a named list with sparse matrices for each chromosome. Each sparse matrix at position i,j contains the number of reads that map to the cell j in the genomic coordinate i (\strong{in 1 based format}).}

\item{\code{path}}{original path from which the object was loaded (optional)}

\item{\code{genomic_bins}}{an intervals set with the genomic bins. Should have columns named "chrom", "start", "end" and "name" with the chromosome, start and end positions and name of the bin.}
}}

