% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\docType{class}
\name{McCounts-class}
\alias{McCounts-class}
\alias{McCounts}
\title{McCounts object}
\description{
a McCounts object is a collection of sparse matrices where rows are genomic coordinates and columns are metacells. It is derived from ScCounts using \code{scc_project_on_mc}. The metacell names are store in the \code{cell_names} slot, and the only additional slot is \code{cell_to_metacell} which contains the mapping from single cell names to metacell names.
}
\section{Slots}{

\describe{
\item{\code{cell_to_metacell}}{}
}}

