% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{split_long_peaks}
\alias{split_long_peaks}
\title{Split long peaks into smaller peaks}
\usage{
split_long_peaks(
  marginal_track,
  peaks,
  target_size = 500,
  max_peak_size = 1000,
  very_long = 5000,
  min_peak_size = NULL
)
}
\arguments{
\item{marginal_track}{Name of the 'misha' track with the marginal coverage. You can create it using \code{import_atac_marginal}.}

\item{peaks}{An intervals set with the peaks to split.}

\item{target_size}{The target size of peaks.}

\item{max_peak_size}{Peaks above this size would be splitted into smaller peaks.}

\item{very_long}{Peaks above this size would be splitted arbitrarily into smaller peaks before fitting the best offset.}

\item{min_peak_size}{Peaks below this size would be discarded.}
}
\description{
Peaks are splitted into smaller peaks if they are longer than \code{target_size}. The splitting is done by first checking
if the peak has a length above \code{very_long}. If it does, the peak is splitted arbitrarily into chunks of size that is approximately \code{very_long}. \cr
Then, the chunks are splitted into smaller chunks of approximately size \code{target_size}. This splitting is done by first finding the
best offset to start from and then splitting the peaks into chunks of size \code{target_size} starting from that offset. \cr
Detection of the best offset is done by substracting the mean of each interval from each coverage, removing values which became
zero, and then correlating thr marginal coverage with simulated 'triangle' peaks starting at different offsets.
}
\examples{
\dontrun{
split_peaks <- split_long_peaks("pbmc_atac.marginal", peaks = peaks, target_size = 500, max_peak_size = 1e3, very_long = 5e3, min_peak_size = 20)
}

}
