% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{subset_peak_clusters}
\alias{subset_peak_clusters}
\title{Subset McPeaks by certain clusters}
\usage{
subset_peak_clusters(
  atac_mc,
  cluster_membership,
  clusters_to_keep,
  reverse = TRUE
)
}
\arguments{
\item{atac_mc}{\itemize{
\item an McPeaks object
}}

\item{cluster_membership}{\itemize{
\item which cluster each peak is a member of
}}

\item{clusters_to_keep}{\itemize{
\item a vector of clusters to keep (or exclude, if \code{reverse == TRUE})
}}

\item{reverse}{(optional) - a logical/flag whether to keep (default - TRUE) or remove the clusters in \code{clusters_to_keep}}
}
\value{
the atac_mc object only with the clusters (peaks) of interest (not saved in the "ignore_..." slots)
}
\description{
Subset McPeaks by certain clusters
}
\examples{
\dontrun{
peak_cl_km <- gen_atac_peak_clust(atac_mc, k = 15)
atac_mc_subset <- subset_peak_clusters(atac_mc, cluster_membership = peak_cl_km, clusters_to_keep = c(4, 5, 8))
}
}
