% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_track_name}
\alias{temp_track_name}
\title{Get a temporary track name}
\usage{
temp_track_name(prefix = "", envir = parent.frame())
}
\arguments{
\item{prefix}{a prefix for the track name}

\item{envir}{environment to to pass the \code{withr::defer} function (do not change unless you know what you are doing)}
}
\value{
a temporary track name
}
\description{
This function returns a temporary track name, which would be deleted when the calling function exits.
}
\examples{
func <- function() {
    tmp <- temp_track()
    print(tmp)
    gtrack.create_sparse(tmp, description = "", intervals = gintervals.all(), values = rep(1, nrow(gintervals.all())))
    gtrack.exists(tmp) # returns TRUE
    a <- gextract(tmp, gintervals.all())
    print(head(a))
    return(tmp)
}
tmp <- func()
gtrack.exists(tmp) # returns FALSE
}
