% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bam.R
\name{write_metacell_cell_names}
\alias{write_metacell_cell_names}
\title{Utility function to write metacell names to files}
\usage{
write_metacell_cell_names(mcatac, c2mc_path = NULL)
}
\arguments{
\item{mcatac}{an McATACPeaks object (from which to get metacell assignments)}

\item{c2mc_path}{(optional) path to write out metacell assignments to}
}
\description{
Utility function to write metacell names to files
}
