% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-counts.R
\name{write_sc_counts_from_10x}
\alias{write_sc_counts_from_10x}
\title{Write single cell counts from 10x 'outs' directory}
\usage{
write_sc_counts_from_10x(
  input_dir,
  out_dir,
  cell_suffix = NULL,
  genome = NULL,
  bin_size = 50000000,
  overwrite = FALSE,
  id = "",
  description = "",
  verbose = FALSE,
  tabix_bin = "tabix",
  chromosomes = NULL
)
}
\arguments{
\item{input_dir}{path to the 10x 'outs' directory}

\item{out_dir}{output directory.}

\item{cell_suffix}{a suffix to add to the cell names (optional). This is useful when you want to merge multiple batches of cells.
The suffix would be added in the following way: \code{paste0(cell_name, "-", cell_suffix)}.}

\item{bin_size}{Size of the genomic bins to use (in bp). Each chromsome will be chunked into bins with size which is
smaller than this value. Default is 50Mb.}

\item{overwrite}{overwrite existing directory (optional)}

\item{id}{an identifier for the object, e.g. "pbmc".}

\item{description}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"}

\item{verbose}{verbose output (optional)}

\item{tabix_bin}{path to the tabix binary (optional)}

\item{chromosomes}{process only specific chromosomes (optional)}
}
\value{
None
}
\description{
The function reads the "outs/atac_fragments.tsv.gz" file, and "outs/filtered_feature_bc_matrix/barcodes.tsv.gz" file and
writes an ScCounts object to "out_dir".
}
\examples{
\dontrun{
write_sc_counts_from_10x("outs", "pbmc_reads", cell_suffix = "batch1", genome = "hg38")
}

}
