% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-raw-counts.R
\name{write_sc_counts_from_bam}
\alias{write_sc_counts_from_bam}
\title{Write single cell counts from an indexed bam file}
\usage{
write_sc_counts_from_bam(
  bam_file,
  out_dir,
  cell_names,
  genome = NULL,
  id = "",
  description = "",
  min_mapq = NULL,
  samtools_bin = "/home/feshap/src/samtools-1.15.1/samtools",
  samtools_opts = NULL,
  num_reads = NULL,
  verbose = TRUE,
  num_cores = parallel::detectCores()
)
}
\arguments{
\item{bam_file}{name of the bam file}

\item{out_dir}{output directory.}

\item{cell_names}{a vector with the cell names or an ScATAC object}

\item{genome}{genome name (e.g. hg19). Will be inferred from the ScATAC object if provided}

\item{id}{an identifier for the object, e.g. "pbmc".}

\item{description}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"}

\item{min_mapq}{minimal mapping quality (optional)}

\item{samtools_bin}{path to samtools executable}

\item{samtools_opts}{additional options for samtools (e.g. "--subsample 0.1")}

\item{num_reads}{number of reads to process (optional)}

\item{verbose}{verbose output (optional)}

\item{num_cores}{number of cores to use (optional)}
}
\value{
None
}
\description{
Write single cell counts from an indexed bam file
}
\examples{
\dontrun{
write_sc_counts_from_bam("pbmc_data/possorted_bam.bam", "pbmc_reads", cell_names = atac_sc)
}

}
