% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr-bias.R
\name{add_const_peaks}
\alias{add_const_peaks}
\title{Add constitutive peaks to McPeaks object}
\usage{
add_const_peaks(mcatac, const_threshold)
}
\arguments{
\item{mcatac}{McPeaks object.}

\item{const_threshold}{The threshold value for determining constitutive peaks. A peak would be considered constitutive if its minimal expression across all cell types is greater than or equal to this threshold.}
}
\value{
The McPeaks object with the added field.
}
\description{
This function adds a field named {.field const} to the {.field peaks} field of the McPeaks object. The field contains a named logical vector indicating whether each peak is constitutive or not.
}
\examples{
\dontrun{
mcatac <- import_from_matrix(mat, peaks, genome = "mm10", class = "McPeaks")
mcatac <- normalize_egc(mcatac, "gastrulation.marginal") #'
mcatac <- add_const_peaks(mcatac, -16)
}

}
