% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_intervals}
\alias{annotate_intervals}
\title{Annotate an intervals set}
\usage{
annotate_intervals(
  intervals,
  MIN_PROXIMAL = 1000,
  MAX_PROXIMAL = 20000,
  MIN_DISTAL = 20000,
  MAX_DISTAL = 1000000,
  EXONIC_PEAK_DIST = 0
)
}
\arguments{
\item{intervals}{the intervals set to annotate}
}
\value{
an intervals set with the following annotations (TODO)
}
\description{
Annotate an intervals set
}
