% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_peaks}
\alias{annotate_peaks}
\title{Annotate ATAC peaks}
\usage{
annotate_peaks(atac)
}
\arguments{
\item{atac}{a McATACPeaks or ScATACPeaks object}
}
\value{
the original \code{atac} object, with the peaks annotated using \code{annotate_intervals}
}
\description{
Annotate ATAC peaks
}
\examples{
\dontrun{
pbmc_atac_mc <- annotate_peaks(pbmc_atac_mc)
}
}
