% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\docType{class}
\name{ATAC-class}
\alias{ATAC-class}
\alias{ATAC}
\title{ATAC objects}
\description{
ATAC is a shallow object holding ATAC data over cells/metacells. Minimally it should include a count matrix of peaks over cells/metacells, \code{PeakIntervals} which hold the coordinates of the peaks and the id of the genome assembly of the peaks. ScATAC and
McATAC extend the ATAC object by adding metadata and additional slots.
}
\section{Slots}{

\describe{
\item{\code{mat}}{a numeric matrix where rows are peaks and columns are cells/metacells. Can be a sparse matrix.}

\item{\code{peaks}}{misha intervals set. Can contain a field named 'peak_name' with a unique name per peak. Both the names and intervals should be unique (a peak cannot appear more than once).}

\item{\code{genome}}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{\code{(optional)}}{metadata data frame with a column called 'metacell' and additional metacell annotations for McATAC, or 'cell_id' and per-cell annotatoins for ScATAC. The constructor can also include or the name of a delimited file which contains such annotations.}
}}

