% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\docType{class}
\name{McATAC-class}
\alias{McATAC-class}
\alias{McATAC}
\alias{ScATAC-class}
\alias{ScATAC}
\title{McATAC}
\description{
An ATAC object with data over metacells

An ATAC object with data over cells
}
