% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATAC.R
\name{atac_ignore_peaks}
\alias{atac_ignore_peaks}
\title{Set ignored (i.e. blacklisted) peaks}
\usage{
atac_ignore_peaks(atac, ig_peaks, reset = FALSE)
}
\arguments{
\item{atac}{an ScATAC or McATAC object}

\item{ig_peaks}{a PeakIntervals object, or vector of peak names to ignore}

\item{reset}{rest the current ignore policy if exists. When set to \code{TRUE}, the current ignore policy will be removed, otherwise the new ignore policy will be appended to the current one.}
}
\value{

}
\description{
Given a list of peaks to ignore, this will remove the given peaks from the {.code ignore_peaks} and {.code ignore_pmat} slots. Note that the matrix and peaks would be reordered.
}
\examples{
\dontrun{
max_peak_length <= 1000
peak_stats <- get_peak_coverage_stats(atac_sc, scale = 100)
too_long_peaks <- peak_stats$peak_name[peak_stats$len > max_peak_length]
atac_sc_filtered <- atac_ignore_peaks(atac_sc, too_long_peaks)
}
}
