% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McATACPeaks.R
\docType{class}
\name{McATACPeaks-class}
\alias{McATACPeaks-class}
\alias{McATACPeaks}
\alias{ScATACPeaks-class}
\alias{ScATACPeaks}
\title{McATACPeaks}
\description{
An ATACPeaks object with data over metacells

An ATACPeaks object with data over cells
}
\section{Slots}{

\describe{
\item{\code{egc}}{normalized metacell accessibility: fraction of accessibility per metacell scaled to the \code{mc_size_eps_q} quantile of
metacell size. Accessibility is normalized by peak length.}

\item{\code{fp}}{a matrix showing for each peak (row) the relative enrichment of umis in log2 scale, i.e. \eqn{log2((1 + egc) / median(1 + egc))}}

\item{\code{mc_size_eps_q}}{quantile of MC size (in UMIs) to scale the number of UMIs per metacell. See \code{project_atac_on_mc}}

\item{\code{rna_egc}}{normalized gene expression per gene per metacell (optional). Can be created using \code{add_mc_rna}}
}}

