% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{call_peaks}
\alias{call_peaks}
\title{Call peaks from ATAC marginals track}
\usage{
call_peaks(
  marginal_track,
  quantile_thresh = 0.9,
  min_umis = 8,
  split_peaks = TRUE,
  target_size = 500,
  max_peak_size = 1000,
  very_long = 5000,
  min_peak_size = 200,
  window_size = target_size,
  genome = NULL,
  seed = 60427
)
}
\arguments{
\item{marginal_track}{Name of the 'misha' track to call peaks from. You can create it using \code{import_atac_marginal}.}

\item{quantile_thresh}{Quantile of the marginal track above which peaks are called.}

\item{min_umis}{Minimal number of UMIs to call a peak.}

\item{split_peaks}{Split peaks that are longer than \code{max_peak_size} into smaller peaks using \code{split_long_peaks}.}

\item{target_size}{The target size of peaks.}

\item{max_peak_size}{Peaks above this size would be splitted into smaller peaks.}

\item{very_long}{Peaks above this size would be splitted arbitrarily into smaller peaks before fitting the best offset.}

\item{min_peak_size}{Peaks below this size would be discarded.}

\item{window_size}{Smoothing window size (from each side of the iterator) to use of the marginal track. Default: traget_size.}

\item{genome}{Genome name, such as 'hg19' or 'mm10'. If NULL - the current misha database is used.}

\item{seed}{random seed for reproducibility (\code{misha::gquantiles} sometimes samples the data)}
}
\value{
an intervals set with the called peaks
}
\description{
Peaks are called by screening for genomic regions with a number of UMIs above a quantile of the marginal ATAC counts,
and above \code{min_umis}. Peaks that are longer than \code{max_peak_size} would be splitted equally into smaller peaks.
}
\examples{
\dontrun{
peaks <- call_peaks("pbmc_atac.marginal", quantile_thresh = 0.9, min_umis = 8, max_peak_size = 600, genome = "hg38")
}

}
