% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bam.R
\name{convert_wigs_to_tracks}
\alias{convert_wigs_to_tracks}
\title{Convert WIG files to misha tracks}
\usage{
convert_wigs_to_tracks(
  wig_dir,
  track_name_prefix = NULL,
  description = NULL,
  parallel = TRUE,
  force = FALSE,
  nc = parallel::detectCores()
)
}
\arguments{
\item{wig_dir}{path to folder containing WIG files}

\item{track_name_prefix}{name to prepend to track.}

\item{description}{(optional) description for misha tracks}

\item{parallel}{(optional) whether to use parallel computation}

\item{force}{(optional) whether to force rewrite of existing misha tracks}

\item{nc}{(optional) number of cores for parallel computations}
}
\value{
error log
}
\description{
Convert WIG files to misha tracks
}
