% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_atac_clust_misha}
\alias{export_atac_clust_misha}
\title{Generate a misha track for each atac metacell cluster}
\usage{
export_atac_clust_misha(mc_atac, mc_atac_clust, clust_names)
}
\arguments{
\item{mc_atac}{McATAC object}

\item{mc_atac_clust}{output of \code{gen_atac_mc_clust}}

\item{clust_names}{names for each metacell cluster}

\item{track_prefix}{prefix for generated misha tracks.}
}
\description{
generate a track for each metacell cluster, of the form \code{track_prefix.name}, where names
are given at \code{clust_names}
}
