% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_atac_clust_ucsc}
\alias{export_atac_clust_ucsc}
\title{Generate a ucsc genome browser file for each metacell cluster}
\usage{
export_atac_clust_ucsc(
  atac_mc,
  track_prefix,
  output_dir = getwd(),
  clust_vec = NULL,
  normalization = "none"
)
}
\arguments{
\item{atac_mc}{McPeaks object}

\item{track_prefix}{prefix for generated misha tracks.}

\item{output_dir}{(optional) name of the directory to write files to}

\item{clust_vec}{(optional) a vector of length #metacells representing an annotation/clustering (can be output of \code{gen_atac_mc_clust})}

\item{normalization}{(optional) normalization method, either 'none', 'lfcom' (log2 fold-change over median), 'zs' (Z-scores)}
}
\description{
Generate a ucsc genome browser file for each metacell cluster
}
