% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_h5ad}
\alias{export_to_h5ad}
\title{Write a McATAC or ScATAC object to an h5ad file}
\usage{
export_to_h5ad(object, out_file)
}
\arguments{
\item{object}{McATAC or ScATAC object}

\item{out_file}{name of the output file}
}
\value{
None.
}
\description{
Write a McATAC or ScATAC object to an h5ad file
}
