% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.r
\name{filter_features}
\alias{filter_features}
\title{Filter features by summary statistics}
\usage{
filter_features(
  scatac,
  minimal_max_umi = NULL,
  min_peak_length = NULL,
  max_peak_length = NULL
)
}
\arguments{
\item{scatac}{the McATAC object to analyze}

\item{minimal_max_umi}{(optional) threshold on minimum maximal coverage - i.e. remove all peaks that DON'T have a cell with at least \code{abs_cov_thresh} UMIs}

\item{min_peak_length}{(optional) remove all peaks that are less than \code{min_cov_thresh} base-pairs long}

\item{max_peak_length}{(optional) remove all peaks that are more than \code{min_cov_thresh} base-pairs long}
}
\value{
scatac - the filtered ScATAC object
}
\description{
Remove ATAC peaks with low coverage, or that are too long etc.
}
\examples{
\dontrun{
## quantiles of peak lengths before filtering
quantile(my_scatac@peaks$end - my_scatac@peaks$start, (0:10) / 10)
my_scatac <- filter_features(scatac = my_scatac, minimal_max_umi = 3, min_peak_length = 200, max_peak_length = 1000)

## quantiles of peak lengths after filtering
quantile(my_scatac@peaks$end - my_scatac@peaks$start, (0:10) / 10)
}
}
