% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.r
\name{fix_missing_chroms_in_peaks}
\alias{fix_missing_chroms_in_peaks}
\title{Utility function that makes fake intervals for chromosome/contig not present in an interval set}
\usage{
fix_missing_chroms_in_peaks(peaks)
}
\arguments{
\item{peaks}{(optional) - a PeakIntervals or regular misha intervals object}
}
\value{
the same object with fake peaks for missing chromosomes
}
\description{
Explanation: as a validation step, gextract requires that every chromosome/contig in the ALLGENOME variable also be present in the track directory, so we add fake sequences in all missing chromosomes/contigs
}
\examples{
\dontrun{
peaks_fix <- fix_missing_chroms_in_peaks(sc_atac@peaks)
}
}
