% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{gen_atac_mc_clust}
\alias{gen_atac_mc_clust}
\title{Cluster metacells based on atac profiles}
\usage{
gen_atac_mc_clust(
  atac_mc,
  k = NULL,
  peak_set = NULL,
  use_prior_annot = TRUE,
  annot = "cell_type"
)
}
\arguments{
\item{atac_mc}{a McATAC object}

\item{k}{number of clusters}

\item{peak_set}{a PeakIntervals object (or a misha intervals set). if NULL - the peaks from \code{atac_mc} would be used.}

\item{use_prior_annot}{when TRUE - use the metacell annotation to generate metacell clusters. Clusters would be generated based on a categorical field \code{annot} from the \code{metadata} slot in the McATAC object.}

\item{annot}{name of the field to use when \code{use_prior_annot} is TRUE.}
}
\description{
Cluster metacells based on atac profiles
}
