% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{gen_atac_peak_clust}
\alias{gen_atac_peak_clust}
\title{Cluster atac peaks based on atac distributions}
\usage{
gen_atac_peak_clust(atac_mc, k, cluster_on = "fp", peak_set = NULL, ...)
}
\arguments{
\item{atac_mc}{a McATAC object}

\item{k}{number of clusters}

\item{cluster_on}{(optional; default - fp) what matrix (\code{mat}/\code{fp}/\code{egc})to cluster on}

\item{peak_set}{\itemize{
\item (optional) a subset of peaks of \code{atac_mc@peaks} on which to cluster
}}

\item{...}{
  Arguments passed on to \code{\link[tglkmeans:TGL_kmeans]{tglkmeans::TGL_kmeans}}
  \describe{
    \item{\code{df}}{data frame. Each row is a single observation and each column is a dimension.
the first column can contain id for each observation (if id_column is TRUE).}
    \item{\code{metric}}{distance metric for kmeans++ seeding. can be 'euclid', 'pearson' or 'spearman'}
    \item{\code{max_iter}}{maximal number of iterations}
    \item{\code{min_delta}}{minimal change in assignments (fraction out of all observations) to continue iterating}
    \item{\code{verbose}}{display algorithm messages}
    \item{\code{keep_log}}{keep algorithm messages in 'log' field}
    \item{\code{id_column}}{\code{df}'s first column contains the observation id}
    \item{\code{reorder_func}}{function to reorder the clusters. operates on each center and orders by the result. e.g. \code{reorder_func = mean} would calculate the mean of each center and then would reorder the clusters accordingly. If \code{reorder_func = hclust} the centers would be ordered by hclust of the euclidian distance of the correlation matrix, i.e. \code{hclust(dist(cor(t(centers))))}
if NULL, no reordering would be done.}
    \item{\code{hclust_intra_clusters}}{run hierarchical clustering within each cluster and return an ordering of the observations.}
    \item{\code{seed}}{seed for the c++ random number generator}
    \item{\code{parallel}}{cluster every cluster parallely (if hclust_intra_clusters is true)}
  }}
}
\value{
a tglkmeans clustering object
}
\description{
Cluster atac peaks based on atac distributions
}
\examples{
\dontrun{
     my_atac_mc <- gen_atac_peak_clust(my_atac_mc, k = 16, cluster_on = 'mat')

     dyn_p <- identify_dynamic_peaks(my_atac_mc)
     my_atac_mc <- gen_atac_peak_clust(my_atac_mc, k = 16, cluster_on = 'fp', peak_set = dyn_p)
}

}
