% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{gen_atac_peak_clust}
\alias{gen_atac_peak_clust}
\title{Cluster atac peaks based on atac distributions}
\usage{
gen_atac_peak_clust(atac_mc, k, peak_set = NULL)
}
\arguments{
\item{atac_mc}{a McATAC object}

\item{k}{number of clusters}

\item{peak_set}{a PeakInervals object (or a misha intervals set). if NULL - he peaks from \code{atac_mc} would be used.}
}
\description{
Cluster atac peaks based on atac distributions
}
