% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.r
\name{gen_random_genome_peak_motif_matrix}
\alias{gen_random_genome_peak_motif_matrix}
\title{Generate random genome motif PSSM matrix}
\usage{
gen_random_genome_peak_motif_matrix(
  num_peaks = 1e+05,
  peak_width = 200,
  bp_from_chrom_edge_to_avoid = 3e+06,
  datasets_of_interest = NULL,
  motif_regex = NULL,
  motif_tracks = NULL,
  parallel = TRUE
)
}
\arguments{
\item{num_peaks}{total number of intervals (will be divided proportionately between chromosomes) in background}

\item{peak_width}{(optional) - size of region around peak centers to extract motif energies for}

\item{bp_from_chrom_edge_to_avoid}{regions (in bp) from edges of chromosomes to avoid sampling from (e.g. avoid acrocentric centromeres)}

\item{datasets_of_interest}{(optional) - names of pssm datasets (\link{name}.key-\link{name}.data file combinations) to calculate PSSM values for}

\item{motif_regex}{(optional) - a vector of regular expressions for which to match motif track names and extract motif PSSMs}

\item{motif_tracks}{(optional) - misha track names for which to extract motif PSSMs}

\item{parallel}{(optional) - whether to use parallel computations}
}
\value{
a matrix of peaks (rows) vs. aggregated motif energies (columns)
}
\description{
Generate random genome motif PSSM matrix
}
\examples{
\dontrun{
     d_vs_rg <- calculate_d_stats(pssm_fg = my_peak_motif_mat, pssm_bg <- random_genome_motif_mat,
                               parallel = F)
}
}
