% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bam.r
\name{generate_per_metacell_bams}
\alias{generate_per_metacell_bams}
\title{Subset 10X atac_possorted.bam file into per-metacell BAM files}
\usage{
generate_per_metacell_bams(bam_path, mcatac, out_dir = NULL, c2mc_path = NULL)
}
\arguments{
\item{bam_path}{path to the 10X atac_possorted.bam file}

\item{mcatac}{an McATAC object (from which to get metacell assignments)}

\item{out_dir}{(optional) directory to output per-metacell bam files}

\item{c2mc_path}{(optional) path to write out metacell assignments to}
}
\description{
Subset 10X atac_possorted.bam file into per-metacell BAM files
}
