% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pheatmap_annotation}
\alias{generate_pheatmap_annotation}
\title{Function to generate generic pheatmap annotation}
\usage{
generate_pheatmap_annotation(
  clust_vec,
  feature_type = NULL,
  feature_annotation = NULL
)
}
\arguments{
\item{clust_vec}{a clustering vector to generically annotate}

\item{feature_type}{(optional) type of the feature that was clustered}

\item{feature_name}{(optional) name of the feature that was clustered}
}
\value{
2-element list containing a pheatmap-compatible dataframe and a list containing a named vector
}
\description{
Generate a generic pheatmap-compatible annotation when clustering unannotated data and plotting with pheatmap
}
