% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bam.R
\name{generate_wigs_from_bams}
\alias{generate_wigs_from_bams}
\title{Write WIGs from BAMs}
\usage{
generate_wigs_from_bams(
  bam_folder_path,
  output_path = file.path(bam_folder_path, "wig_output"),
  parallel = TRUE,
  nc = parallel::detectCores()
)
}
\arguments{
\item{bam_folder_path}{path to BAMs}

\item{output_path}{(optional) where to put WIGs. Default - bam_folder_path/wig_output}

\item{parallel}{(optional) whether to do parallel computations}

\item{nc}{(optional) number of cores for parallel computations}
}
\value{
error log
}
\description{
Write WIGs from BAMs
}
