% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{get_rna_egc}
\alias{get_rna_egc}
\title{Get the normalized RNA expression matrix (egc) from a McATAC object}
\usage{
get_rna_egc(atac_mc, genes = NULL, rm_zeros = TRUE, epsilon = 0.00001)
}
\arguments{
\item{atac_mc}{a McATAC object with RNA expression (using \code{add_mc_rna})}

\item{genes}{list of genes to match. Default (NULL): all genes}

\item{rm_zeros}{remove genes with no RNA expression in any metacell. Default: TRUE}

\item{epsilon}{regularization factor added to the log normalized expression}
}
\value{
a matrix with normalized RNA expression values for each gene (rows) and metacell (columns)
}
\description{
Get RNA expression data, normalized by the total RNA expression in each metacell.
}
\examples{
\dontrun{
rna_mat <- get_rna_egc(atac_mc)
rna_mat <- get_rna_egc(atac_mc, genes = c("MESP1", "MESP2", "PF4"))
rna_mat <- get_rna_egc(atac_mc, rm_zeros = FALSE)
}

}
