% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{get_rna_marker_matrix}
\alias{get_rna_marker_matrix}
\title{Get enrichment matrix for marker genes}
\usage{
get_rna_marker_matrix(
  atac_mc,
  markers = NULL,
  force_cell_type = TRUE,
  rm_zeros = TRUE,
  epsilon = 0.00001,
  ...
)
}
\arguments{
\item{atac_mc}{a McATACPeaks object with RNA expression (using \code{add_mc_rna})}

\item{markers}{a list of marker genes. If NULL - the function uses \code{get_rna_markers} with default parameters which can be overridden
using the ellipsis \code{...}.}

\item{force_cell_type}{do not split cell types when ordering the metacells. Default: TRUE}

\item{rm_zeros}{remove genes with no RNA expression in any metacell. Default: TRUE}

\item{epsilon}{regularization factor added to the log normalized expression}

\item{...}{
  Arguments passed on to \code{\link[=get_rna_markers]{get_rna_markers}}
  \describe{
    \item{\code{n_genes}}{maximal number of genes to return. Default: 100}
    \item{\code{minimal_max_log_fraction}}{take only genes with at least one value
(in log fraction units - normalized egc) above this threshold}
    \item{\code{minimal_relative_log_fraction}}{take only genes with at least one value with relative
log fraction (mc_fp) above this this value}
    \item{\code{fold_change_reg}}{regularization factor for the fold change calculation (fold_change would be changed to
\code{fold_change = fold_change + fold_change_reg})}
    \item{\code{genes}}{list of genes to match. Default (NULL): all genes}
  }}
}
\value{
A matrix with log2 normalized counts of gene expression for each marker gene (rows) and metacell (columns). \cr
The columns are columns are clustered using \code{hclust} with "ward" linkage, on the euclidean distance
of the pearson correlation matrix between the columns. Then, the gene with the highest number of metacells with fold change (abs) above 1 (gene1) is selected together with the gene most anti-correlated to it (gene2), and the \code{hclust} is ordered using the difference (gene1 - gene2). \cr
If \code{force_cell_type} is TRUE and \code{atac_mc@metadata} has a field named "cell_type", the columns are only ordered only within each
cell type.
}
\description{
Get enrichment matrix for marker genes
}
\examples{
\dontrun{
marker_mat <- get_rna_marker_matrix(atac_mc)
marker_mat <- get_rna_marker_matrix(atac_mc, n_genes = 100)
}

}
