% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{has_rna}
\alias{has_rna}
\title{Does the McPeaks object contain per-metacell gene expression data?}
\usage{
has_rna(atac_mc)
}
\arguments{
\item{atac_mc}{an McPeaks object}
}
\value{
TRUE if the McPeaks object contains per-metacell gene expression data, FALSE otherwise
}
\description{
Does the McPeaks object contain per-metacell gene expression data?
}
\examples{
\dontrun{
has_rna(atac_mc)
}

}
