% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_from_10x}
\alias{import_from_10x}
\title{Create an ScATAC object from 10x directory}
\usage{
import_from_10x(
  dir,
  genome,
  id = NULL,
  description = NULL,
  metadata = NULL,
  matrix_fn = file.path(dir, "matrix.mtx"),
  cells_fn = file.path(dir, "barcodes.tsv"),
  features_fn = file.path(dir, "features.tsv")
)
}
\arguments{
\item{dir}{10x directory. Should have the following files: 'matrix.mtx', 'barcodes.tsv' and 'features.tsv'.
See 'Cell Ranger Arc' for more details: https://support.10xgenomics.com/single-cell-multiome-atac-gex/software/overview/welcome}

\item{genome}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{id}{an identifier for the object, e.g. "pbmc". If NULL - a random id would be assigned.}

\item{description}{(Optional) description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)".}

\item{metadata}{(Optional) data frame with a column called 'metacell' and additional metacell annotations, or the name of a delimited file which contains such annotations.}

\item{matrix_fn}{if \code{dir} is missing, the filename of the matrix to import ("matrix.mtx" or "matrix.mtx.gz")}

\item{cells_fn}{if \code{dir} is missing, the filename of the cells to import ("barcodes.tsv" or "barcodes.tsv.gz")}

\item{features_fn}{if \code{dir} is missing, the filename of the features to import ("features.tsv" or "features.tsv.gz")}
}
\value{
an ScATAC object
}
\description{
Create an ScATAC object from 10x directory
}
\examples{
\dontrun{
atac <- import_from_10x("./pbmc_data", genome = "hg38")
atac
}

}
