% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_from_10x}
\alias{import_from_10x}
\title{Create an ScATAC object from 10x directory}
\usage{
import_from_10x(
  dir = NULL,
  matrix_fn = file.path(dir, "matrix.mtx"),
  cells_fn = file.path(dir, "barcodes.tsv"),
  features_fn = file.path(dir, "features.tsv")
)
}
\arguments{
\item{dir}{10x directory. Should have the following files: 'matrix.mtx', 'barcodes.tsv' and 'features.tsv'.
See 'Cell Ranger Arc' for more details: https://support.10xgenomics.com/single-cell-multiome-atac-gex/software/overview/welcome}

\item{matrix_fn}{if \code{dir} is missing, the filename of the matrix to import ("matrix.mtx")}

\item{cells_fn}{if \code{dir} is missing, the filename of the cells to import ("barcodes.tsv")}

\item{features_fn}{if \code{dir} is missing, the filename of the features to import ("features.tsv")}
}
\value{
an ScATAC object
}
\description{
Create an ScATAC object from 10x directory
}
\examples{
\dontrun{
atac <- import_from_10x("./pbmc_data")
atac
}

}
