% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_from_h5ad}
\alias{import_from_h5ad}
\title{Create an ScATAC object from an h5ad file}
\usage{
import_from_h5ad(file)
}
\arguments{
\item{file}{name of an h5ad file with ATAC data}
}
\value{
an ScATAC object
}
\description{
Create an ScATAC object from an h5ad file
}
