% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcATAC-package.R
\docType{package}
\name{mcATAC-package}
\alias{mcATAC}
\alias{mcATAC-package}
\title{mcATAC: Metacell analysis for ATAC data}
\description{
Metacell analysis for ATAC data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tanaylab/mcATAC}
  \item Report bugs at \url{https://github.com/tanaylab/mcATAC/issues}
}

}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

Authors:
\itemize{
  \item Yonatan Shapira \email{yonatan.shapira@weizmann.ac.il}
}

}
\keyword{internal}
