% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{mcc_metacell_total_cov}
\alias{mcc_metacell_total_cov}
\title{Return the total coverage of each metacell in an McCounts object}
\usage{
mcc_metacell_total_cov(mc_counts, metacells = NULL)
}
\arguments{
\item{mc_counts}{a McCounts object}

\item{metacells}{names of metacells to include. Default: all metacells.}
}
\value{
a named vector with the total coverage for each metacell
}
\description{
Return the total coverage of each metacell in an McCounts object
}
\examples{
\dontrun{
mc_counts <- mcc_read("pbmc_reads_mc")
mc_covs <- mcc_cov(mc_counts)
}

}
