% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCounts.R
\name{mcc_to_tracks}
\alias{mcc_to_tracks}
\title{Create a misha track for each metacell in a McCounts object}
\usage{
mcc_to_tracks(
  mc_counts,
  track_prefix,
  metacells = NULL,
  num_cores = parallel::detectCores()
)
}
\arguments{
\item{mc_counts}{A McCounts object}

\item{track_prefix}{The prefix of the tracks to create. Track names will be of the form "{track_prefix}.mc{metacell}"}

\item{metacells}{metacells for which to create tracks. If NULL, all metacells will be used.}

\item{num_cores}{The number of cores to use.}
}
\value{
None.
}
\description{
This would create a sparse track of the form "{track_prefix}.mc{metacell}" for each metacell in the McCounts object.
}
\examples{
\dontrun{
mcc_to_tracks(mc_counts, "pbmc_mc")
}

}
