% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACTracks-cache.R
\name{mct_get_mat}
\alias{mct_get_mat}
\title{Get an ATAC matrix from a McTracks object}
\usage{
mct_get_mat(
  mct,
  intervals,
  downsample = FALSE,
  downsample_n = NULL,
  force = FALSE,
  seed = NULL
)
}
\arguments{
\item{mct}{a McTrack object}

\item{downsample}{return a downsampled matrix. See description.}

\item{downsample_n}{total coverage goal. See description. Default: lower 5\% percentile of the total coverage)}

\item{force}{force the computation of the matrix. If FALSE, the matrix is
retrieved from the cache if it exists.}

\item{seed}{random seed for the downsampling.}

\item{intervlas}{an intervals set}
}
\value{
a matrix where rows are genomic coordinates in the resolution
of \code{mct@resolution} and columns are the metacells in \code{mct@metacells}
}
\description{
Retruns a matrix of genomic coordinates over the metacells. The matrix is cached in memory, so subsequent calls to this function will not require the extraction of the reads from the underlying tracks. \cr
When \code{downsample} is TRUE, downsampling is done by first setting a coverage goal,
then the reads within the region are subsampled relative to the ratio of the metacell coverage and the goal, i.e. \code{N_i * C_i/goal)} where \code{C_i} is the total number of reads in metacell i, \code{N_i} is the total numbers of reads in the region for metacell i, and \code{goal} is the coverage goal. \cr
For example, if the goal is 2M reads, a metacell has a total of 1M reads and 50K reads within the region, the reads within the region are subsampled to \code{50K * 1M/2M = 25K} (50\%). Metacells with less than the goal are removed.
}
