% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-region.R
\name{mct_plot_region}
\alias{mct_plot_region}
\title{Plot a genomic region}
\usage{
mct_plot_region(
  mct,
  intervals,
  downsample = TRUE,
  downsample_n = NULL,
  metacells = NULL,
  colors = c("white", "gray", "black", "gold", "gold")
)
}
\arguments{
\item{mct}{an McTracks object.}

\item{intervals}{an intervals set with the genomic region to plot (a data frame with a single line). Note that if the start or end coordinates are not divisible by the resolution, the region will be extended to the next resolution interval.}

\item{downsample}{return a downsampled matrix. See description.}

\item{downsample_n}{total coverage goal. See description. Default: lower 5th percentile of the total coverage)}

\item{metacells}{a vector of metacells to plot. If NULL, all metacells will be plotted.}

\item{colors}{color pallette for the ATAC signal}
}
\value{
None. The plot is displayed in a new device.
}
\description{
Plot a genomic region
}
\examples{
\dontrun{
mct_plot_region(mct, intervals = data.frame(chr = "chr5", start = 54974252, end = 55074253))
}

}
