% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-region.R
\name{mct_plot_region}
\alias{mct_plot_region}
\title{Plot a genomic region}
\usage{
mct_plot_region(
  mct,
  intervals,
  detect_dca = FALSE,
  downsample = TRUE,
  downsample_n = NULL,
  metacells = NULL,
  colors = c("white", "gray", "black", "gold", "gold"),
  hc = NULL,
  force_cell_type = TRUE,
  gene_annot = FALSE,
  n_smooth = 20,
  n_pixels = 1000,
  ...
)
}
\arguments{
\item{mct}{an McTracks object.}

\item{intervals}{an intervals set with the genomic region to plot (a data frame with a single line). Note that if the start or end coordinates are not divisible by the resolution, the region will be extended to the next resolution interval.}

\item{detect_dca}{mark regions with differential cluster accessibility (DCA)}

\item{downsample}{return a downsampled matrix. See description.}

\item{downsample_n}{total coverage goal. See description. Default: lower 5th percentile of the total coverage)}

\item{metacells}{a vector of metacells to plot. If NULL, all metacells will be plotted.}

\item{colors}{color pallette for the ATAC signal}

\item{hc}{an hclust object to use for the hierarchical clustering of the metacells when \code{detect_dcs=TRUE}. If NULL, metacells will be clustered using \code{mc_hclust_rna}.}

\item{force_cell_type}{do not split cell types when ordering the metacells. Default: TRUE}

\item{gene_annot}{(optional) whether to add gene annotations; these annotations rely on the existence of an \code{annots/refGene.txt} file in the genome's misha directory, and on the existence of an intervals set called "intervs.global.tss" in the genome's misha directory.}

\item{n_smooth}{number of genomic bins to use for smoothing the signal. Signal is smoothed by a rolling sum for each metacell. Default is 20.}

\item{n_pixels}{number of pixels in the plot. The DCA regions would be extended by \code{ceiling(2 * nrow(mat) / n_pixels)}.}

\item{...}{
  Arguments passed on to \code{\link[=mct_diff_access_on_hc]{mct_diff_access_on_hc}}
  \describe{
    \item{\code{mat}}{a matrix where rows are coordinates and columns are metacells}
    \item{\code{sz_frac_for_peak}}{maximal fraction of metacells in a peak or trough. Default is 0.25.}
    \item{\code{peak_lf_thresh1,peak_lf_thresh2,trough_lf_thresh1,trough_lf_thresh2}}{thresholds for the log fold change of the peaks and troughs}
    \item{\code{u_reg}}{regularization factor}
  }}
}
\value{
None. The plot is displayed in a new device.
}
\description{
Plot a genomic region
}
\examples{
\dontrun{
mct_plot_region(mct, intervals = data.frame(chr = "chr5", start = 54974252, end = 55074253))
}

}
