% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bam.R
\name{merge_metacell_bams}
\alias{merge_metacell_bams}
\title{Merge BAMs by metacells}
\usage{
merge_metacell_bams(
  bam_path,
  output_filename,
  mcs,
  parallel = getOption("mcatac.parallel"),
  nc = getOption("mcatac.parallel.nc")
)
}
\arguments{
\item{bam_path}{path to the per-metacell BAM tracks}

\item{output_filename}{what to call the output BAM file}

\item{mcs}{vector of mcs to merge BAMs of}

\item{parallel}{(optional) whether to do parallel computations}

\item{nc}{(optional) number of cores for parallel computations}
}
\description{
Merge BAMs by metacells
}
