% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{name_enhancers}
\alias{name_enhancers}
\title{Name ATAC peaks from TAD names}
\usage{
name_enhancers(atac)
}
\arguments{
\item{atac}{a McATAC or ScATAC or PeakIntervals object}
}
\value{
the original \code{atac} object, with an added field \code{enh_name} to the relevant PeakIntervals dataframe
}
\description{
Name ATAC peaks from TAD names
}
\examples{
\dontrun{
     atac_mc <- name_enhancers(atac_mc)
     my_peaks <- name_enhancers(my_peaks)
}
}
