% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{name_enhancers}
\alias{name_enhancers}
\title{Name ATAC peaks from TAD names}
\usage{
name_enhancers(atac, tad_names = gintervals.load("intervs.global.tad_names"))
}
\arguments{
\item{atac}{a McATAC or ScATAC a PeakIntervals object or a misha intervals set}
}
\value{
the original \code{atac} object, with the field \code{peak_name} replaced by the new TAD-referenced name
}
\description{
The peaks are named in the following pattern "{TAD_name}\emph{{distance_from_TAD_start_in_kbp}".
For the TAD naming scheme see: https://github.com/tanaylab/tadbris. \cr
When two peaks are within the same kbp of each other, they are appended a number to the name, according to
their order on the genome (except for the first peak), i.e.: "{TAD_name}}{distance_from_TAD_start_in_kbp}.{number}".
See also \code{make.unique}.
}
\examples{
\dontrun{
atac_mc <- name_enhancers(atac_mc)
my_peaks <- name_enhancers(my_peaks)
}
}
