% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakIntervals.R
\name{PeakIntervals}
\alias{PeakIntervals}
\title{Construct a PeakIntervals object from misha intervals}
\usage{
PeakIntervals(intervals)
}
\arguments{
\item{intervals}{misha intervals (a data frame which has the following columns: 'chrom', 'start' and 'end')}
}
\value{
a PeakIntervals object
}
\description{
a PeakIntervals object is simply a misha intervals set which can have, optionally, a field called 'peak_name'
which holds the name of the peaks.
}
