% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_atac_peak_map}
\alias{plot_atac_peak_map}
\title{Plot normalized accessibility of peaks over metacells, ordered by clustering}
\usage{
plot_atac_peak_map(
  mc_atac,
  mc_atac_clust,
  peak_clust,
  filepath = NULL,
  dev = NULL,
  clrs = colorRampPalette(c("blue4", "white", "red4"))(100),
  seed = 1337
)
}
\arguments{
\item{mc_atac}{McATAC object}

\item{mc_atac_clust}{output of \code{gen_atac_mc_clust}}

\item{peak_clust}{output of \code{gen_atac_peak_clust}}

\item{filepath}{(optional) path and filename of where to save the figure; if unspecified, figure isn't saved}

\item{dev}{(optional; default - png) graphical device to save figure with}
}
\value{
a pheatmap figure
}
\description{
Plot normalized accessibility of peaks over metacells, ordered by clustering
}
\examples{
\dontrun{
    peak_clust <- gen_atac_peak_clust(my_mcatac, 16)
    plot_atac_peak_map(my_mcatac, my_mcatac@metadata, peak_clust, './my_figures/my_mcatac_heatmap.png')
}
}
