% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_atac_peak_map}
\alias{plot_atac_peak_map}
\title{Plot normalized accessibility of peaks over metacells, ordered by clustering}
\usage{
plot_atac_peak_map(mc_atac, mc_atac_clust, peak_clust)
}
\arguments{
\item{mc_atac}{McATAC object}

\item{mc_atac_clust}{output of \code{gen_atac_mc_clust}}

\item{peak_clust}{output of \code{gen_atac_peak_clust}}
}
\description{
Plot normalized accessibility of peaks over metacells, ordered by clustering
}
\examples{
\dontrun{

}
}
