% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_atac_peak_map}
\alias{plot_atac_peak_map}
\title{Plot normalized accessibility of peaks over metacells, ordered by clustering}
\usage{
plot_atac_peak_map(
  mc_atac,
  mc_atac_clust,
  filename,
  peak_clust = NULL,
  eps_q = 0.1,
  ...
)
}
\arguments{
\item{mc_atac}{McATAC object}

\item{mc_atac_clust}{a clustering of metacells (e.g. output of {.code gen_atac_mc_clust})}

\item{filename}{file name of saved image}

\item{peak_clust}{(optional) output of \code{gen_atac_peak_clust}}

\item{...}{
  Arguments passed on to \code{\link[=save_pheatmap_png]{save_pheatmap_png}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Plot normalized accessibility of peaks over metacells, ordered by clustering
}
\examples{
\dontrun{

}
}
