% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_atac_rna}
\alias{plot_atac_rna}
\title{Plot a scatter of gene expression vs an atac profile}
\usage{
plot_atac_rna(
  mc_atac,
  gene,
  mc_rna = NULL,
  peak = NULL,
  max_dist_to_promoter_peak = 500,
  normalize_atac = TRUE,
  eps_rna = 0.00001,
  tss_intervals = "intervs.global.tss"
)
}
\arguments{
\item{mc_atac}{a McATAC object}

\item{gene}{name of the gene to plot}

\item{mc_rna}{a \code{metacell1} 'mc' object or a \code{metacells} metacell UMI matrix (a matrix where each row is a gene and each column is a metacell). Can be NULL if \code{mc_atac} already contains the gene expression data (added by \code{add_mc_rna}).}

\item{peak}{name of the peak to plot. If NULL - the promoter of \code{gene} would be shown. You can get the peak names from \code{atac@peaks$peak_name}}

\item{max_dist_to_promoter_peak}{how far from \code{gene}'s TSS to search for a promoter-proximal peak}

\item{normalize_atac}{whether to use normalized atac profiles (default: TRUE)}

\item{eps_rna}{added regularization when calculating log expression (Default: 1e-5). Promoter ATAC signal has been observed empirically to often be linearly correlated with log of gene expression.}
}
\value{
a ggplot object with the scatter plot
}
\description{
Plot a scatter of gene expression vs an atac profile
}
\examples{
\dontrun{
p1 <- plot_atac_rna(mc_atac = pbmc_atac_mc, mc_rna = pbmc_rna_mc, gene = "CD4")

## Plot gene vs some peak of interest; here - expression of CD4 vs. promoter ATAC signal of CD8
tss <- gintervals.load("intervs.global.tss")
cd8_tss <- tss[which(tss$geneSymbol == "CD8")[[1]], ]
peaks_near_cd8_tss <- misha::gintervals.neighbors(pbmc_atac_mc@peaks, cd8_tss, mindist = -2e+3, maxdist = 2e+3)
nearest_peak <- peaks_near_cd8_tss[which.min(peaks_near_cd8_tss$dist), ]
p2 <- plot_atac_rna(mc_atac = pbmc_atac_mc, mc_rna = pbmc_rna_mc, gene = "CD4", peak = nearest_peak[, "peak_name"])
}
}
