% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{plot_peak_coverage_density}
\alias{plot_peak_coverage_density}
\title{Plot the coverage density of the peaks}
\usage{
plot_peak_coverage_density(atac, scale = 100, pointsize = 1.5, ...)
}
\arguments{
\item{atac}{an ScATAC or McATAC object}

\item{scale}{for the 'cov_density' field, calculate the number of UMIs per \code{scale}
bp (default: 100)}

\item{pointsize}{Radius of rasterized point. Use \code{0} for single pixels (fastest).}

\item{...}{
  Arguments passed on to \code{\link[scattermore:geom_scattermore]{scattermore::geom_scattermore}}
  \describe{
    \item{\code{mapping}}{passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}}
    \item{\code{data}}{passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}}
    \item{\code{stat}}{passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}}
    \item{\code{position}}{passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}}
    \item{\code{inherit.aes}}{passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}}
    \item{\code{show.legend}}{passed to \code{\link[ggplot2:layer]{ggplot2::layer()}}}
    \item{\code{na.rm}}{Remove NA values, just as with \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}
    \item{\code{interpolate}}{Default FALSE, passed to \code{\link[grid:grid.raster]{grid::rasterGrob()}}.}
    \item{\code{pixels}}{Vector with X and Y resolution of the raster, default \code{c(512,512)}.}
  }}
}
\description{
For each peak, plot a scatter of the peak length vs the number
of UMIs per \code{scale} bp.
}
\examples{
\dontrun{
plot_peak_coverage_density(atac)
}

}
