% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-region.R
\name{plot_region_mat}
\alias{plot_region_mat}
\title{Plot a region of the metacells given a matrix}
\usage{
plot_region_mat(
  mat,
  mc_colors = NULL,
  colors = c("white", "gray", "black", "gold", "gold"),
  intervals = NULL
)
}
\arguments{
\item{mat}{a matrix where rows are coordinates and columns are metacells}

\item{mc_colors}{a vector of colors for the metacells (optional)}

\item{colors}{color pallette for the ATAC signal}

\item{intervals}{the plotted intervals (optional)}
}
\description{
Plot a region of the metacells given a matrix
}
