% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-region.R
\name{plot_region_mat}
\alias{plot_region_mat}
\title{Plot a genomic region given a matrix}
\usage{
plot_region_mat(
  mat,
  mc_colors = NULL,
  colors = c("white", "gray", "black", "gold", "gold"),
  intervals = NULL,
  dac_mat = NULL,
  n_pixels = 1000
)
}
\arguments{
\item{mat}{a matrix where rows are coordinates and columns are metacells}

\item{mc_colors}{a vector of colors for the metacells (optional)}

\item{colors}{color pallette for the ATAC signal}

\item{intervals}{the plotted intervals (optional)}

\item{dac_mat}{a matrix with the differential accessibility (DAC) for the plotted regions (optional). Output of \code{mct_diff_access_on_hc}.}

\item{n_pixels}{number of pixels in the plot. The DAC regions would be extended by \code{ceiling(2 * nrow(mat) / n_pixels)}.}
}
\description{
Plot a genomic region given a matrix
}
