% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-region.R
\name{plot_region_mat}
\alias{plot_region_mat}
\title{Plot a genomic region given a matrix}
\usage{
plot_region_mat(
  mat,
  mc_colors = NULL,
  colors = c("white", "gray", "black", "gold"),
  color_breaks = c(0, 6, 12, 18, 24),
  intervals = NULL,
  resolution = NULL,
  dca_mat = NULL,
  y_seps = NULL,
  y_seps_lty = 2,
  y_seps_lwd = 1,
  n_smooth = 10,
  n_pixels = 1000,
  gene_annot = FALSE,
  genome = NULL,
  plot_x_axis_ticks = TRUE,
  gene_annot_pos = "top",
  flip = FALSE,
  roll_mean = FALSE
)
}
\arguments{
\item{mat}{a matrix where rows are coordinates and columns are metacells}

\item{mc_colors}{a vector of colors for the metacells (optional)}

\item{colors}{color pallette for the ATAC signal}

\item{color_breaks}{a vector of breaks for the color palette}

\item{intervals}{the plotted intervals (optional)}

\item{resolution}{the resolution of the plotted intervals (optional)}

\item{dca_mat}{a matrix with the differential cluster accessibility (DCA) for the plotted regions (optional). Output of \code{mct_diff_access_on_hc}.}

\item{n_smooth}{number of genomic bins to use for smoothing the signal. Signal is smoothed by a rolling sum for each metacell (optional). Default is 20.}

\item{n_pixels}{number of pixels in the plot. The DCA regions would be extended by \code{ceiling(2 * nrow(mat) / n_pixels)} (optional).}

\item{gene_annot}{whether to add gene annotations; these annotations rely on the existence of the existence of an intervals set called "intervs.global.tss" and "intervs.global.exon" in the genome's misha directory. (optional)}

\item{genome}{the genome to use for the gene annotations (optional)}

\item{gene_annot_pos}{the position of the gene annotations ("top" or "bottom")}

\item{flip}{whether to flip the coordinates (optional)}
}
\description{
Plot a genomic region given a matrix
}
