% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_on_mc.R
\name{project_atac_on_mc}
\alias{project_atac_on_mc}
\alias{project_atac_on_mc_from_metacell1}
\alias{project_atac_on_mc_from_h5ad}
\title{Given metacells (usually from RNA data), project ATAC counts to get a McATACPeaks object}
\usage{
project_atac_on_mc(
  atac,
  cell_to_metacell = NULL,
  metadata = NULL,
  min_int_frac = 0.5,
  mc_size_eps_q = 0.1,
  id = NULL,
  description = NULL,
  rm_zero_peaks = TRUE,
  ignore_metacells = -1
)

project_atac_on_mc_from_metacell1(
  atac,
  scdb,
  mc_id,
  metadata = NULL,
  id = NULL,
  description = NULL
)

project_atac_on_mc_from_h5ad(
  atac,
  h5ad_file,
  min_int_frac = 0.5,
  metadata = NULL,
  id = NULL,
  description = NULL
)
}
\arguments{
\item{atac}{an ScATACPeaks object}

\item{cell_to_metacell}{a data frame with a column named "cell_id" with cell id and another column named "metacell" with the metacell the cell is
part of.}

\item{metadata}{(optional) per-metacell metadata. A data frame with a column called 'metacell' and additional metacell annotations.}

\item{min_int_frac}{(optional) minimal expected fraction of intersection of barcodes (cell names) in ScATACPeaks}

\item{mc_size_eps_q}{(optional) quantile of MC size (in UMIs) to scale the number of UMIs per metacell. \eqn{egc_ij} would then be
the fraction of peak i in metacell j multiplied by the \code{mc_size_eps_q} quantile of metacell sizes.}

\item{id}{an identifier for the object, e.g. "pbmc". If NULL - the id would be taken from the scATAC object \code{atac}.}

\item{description}{an identifier for the object, e.g. "pbmc". If NULL - the description would be taken from the scATAC object \code{atac}}

\item{rm_zero_peaks}{remove peaks without any reads (all-zero peaks). Default: TRUE}

\item{ignore_metacells}{a vector of metacells to ignore. Default: \link{-1} (the "outliers" metacell in the metacell2 python package).}

\item{scdb}{a metacell1 \code{scdb} path}

\item{mc_id}{id of the metacell object within \code{scdb}}

\item{h5ad_file}{name of an h5ad file which is the output of 'metacells' python package.}
}
\value{
an McATACPeaks object
}
\description{
Given cell to metacell association, summarise atac counts to generate a McATACPeaks object.
\code{project_atac_on_mc_from_metacell1} projects given a metacell1 'mc' object, while \code{project_atac_on_mc_from_h5ad}
uses the output of the 'metacells' python package (metacell2).
}
\examples{
\dontrun{
atac_sc <- import_from_10x("pbmc_data", genome = "hg38")
data(cell_to_metacell_pbmc_example)
atac_mc <- project_atac_on_mc(atac_sc, cell_to_metacell_pbmc_example)
}

}
