% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCounts.R
\name{read_sc_counts}
\alias{read_sc_counts}
\title{Read an ScCounts object from a directory}
\usage{
read_sc_counts(path, id = NULL, description = NULL)
}
\arguments{
\item{path}{path to the directory containing the object (which was created by \code{write_sc_counts})}

\item{id}{an identifier for the object (optional)}

\item{description}{description of the object (optional)}
}
\value{
an ScCounts object
}
\description{
Read an ScCounts object from a directory
}
\examples{
\dontrun{
counts <- read_sc_counts("pbmc_reads")
}

}
