% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_app}
\alias{run_app}
\title{Run a shiny app for viewing the MCT object}
\usage{
run_app(
  mct,
  hc = NULL,
  mct2 = NULL,
  hc2 = NULL,
  port = NULL,
  host = NULL,
  launch.browser = FALSE,
  chain = NULL,
  chain2 = NULL,
  annotations1 = NULL,
  annotations2 = NULL
)
}
\arguments{
\item{mct}{MCT object}

\item{hc}{an hclust object with clustering of the metacells (optional, see \code{mct_plot_region}}

\item{mct2}{MCT object from a different genome (optional)}

\item{hc2}{an hclust object with clustering of the metacells for the second genome (optional, see \code{mct_plot_region}}

\item{chain}{filename of a chain translating from the genome of mct to the genome of mct2, or an rtracklayer chain object (optional)}

\item{chain2}{filename of a chain translating from the genome of mct2 to the genome of mct, or an rtracklayer chain object (optional)}

\item{annotations1}{an intervals object with annotations for the first genome (optional)}

\item{annotations2}{an intervals object with annotations for the second genome (optional)}
}
\description{
Run a shiny app for viewing the MCT object
}
\examples{
\dontrun{
run_app(mct)
}

}
