% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_app}
\alias{run_app}
\title{Run a shiny app for viewing the MCT object}
\usage{
run_app(mct, port = NULL, host = NULL, launch.browser = FALSE)
}
\arguments{
\item{mct}{MCT object}
}
\description{
Run a shiny app for viewing the MCT object
}
