% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCounts.R
\name{scc_write}
\alias{scc_write}
\alias{mcc_write}
\title{Write a counts object to a file}
\usage{
scc_write(object, out_dir, overwrite = FALSE)

mcc_write(object, out_dir, overwrite = FALSE)
}
\arguments{
\item{object}{a counts object (ScCounts or McCounts)}

\item{out_dir}{the output directory}

\item{overwrite}{whether to overwrite existing directory}
}
\description{
Write a counts object to a file
}
\examples{
\dontrun{
scc_write(sc_counts, "sc_counts")
mcc_write(mc_counts, "mc_counts")
}

}
