% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCounts.R
\docType{class}
\name{ScCounts-class}
\alias{ScCounts-class}
\alias{ScCounts}
\title{ScCounts object}
\description{
ScCounts is a data structure that contains the counts of reads for a set of cells in a set of chromosomes.
}
\section{Slots}{

\describe{
\item{\code{id}}{an identifier for the object, e.g. "pbmc".}

\item{\code{description}}{description of the object, e.g. "PBMC from a healthy donor - granulocytes removed through cell sorting (10k)"}

\item{\code{genome}}{genome assembly of the peaks. e.g. "hg38", "hg19", "mm9", "mm10"}

\item{\code{data}}{a named list with sparse matrices for each chromosome.}

\item{\code{path}}{original path from which the object was loaded (optional)}
}}

