% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-raw-counts.R
\name{write_sparse_matrix_from_bam}
\alias{write_sparse_matrix_from_bam}
\title{Write a sparse-matrix file from an indexed bam file}
\usage{
write_sparse_matrix_from_bam(
  bam_file,
  out_file,
  cell_names,
  region,
  genome = NULL,
  min_mapq = NULL,
  samtools_bin = "samtools",
  samtools_opts = NULL,
  num_reads = NULL,
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{bam_file}{name of the bam file}

\item{out_file}{name of the output file. A ".gz" extension will be added to the file name.}

\item{cell_names}{a vector with the cell names or an ATAC object}

\item{region}{intervals set to filter. See samtools docs \url{http://www.htslib.org/doc/samtools-view.html} for details}

\item{genome}{genome name (e.g. hg19). Will be inferred from the ScATAC object if provided}

\item{min_mapq}{minimal mapping quality (optional)}

\item{samtools_bin}{path to samtools executable}

\item{samtools_opts}{additional options for samtools (e.g. "--subsample 0.1")}

\item{num_reads}{number of reads (within the \code{region}) to process (optional).}

\item{verbose}{verbose output (optional)}

\item{overwrite}{overwrite existing files (optional)}
}
\value{
None
}
\description{
This function reads a bam file and writes a sparse-matrix file, where rows are the genomic coordinates
and columns are the cells. It does so by:
\enumerate{
\item filtering the bam file by the given region (e.g. chromosom, using the "regions" paramter)
\item extracting the cell name using the "CB" tag added by cellranger, while excluding reads which were marked as PCR duplicates (1024)
\item use awk to extract the start coordinates and the tag
\item use unix "sort" and "uniq" to count the number of times each coordinate appears for each cell
\item use an R script to convert the cell names to indices
\item write the sparse-matrix file (matrix-market format) and zip it
} \cr
Note (1): In order to use this function, "samtools" (>= 1.15), "awk", "sed", "sort" and "uniq" must be available at the unix command line.
Note (2): The function takes a while to run - around 13 minutes using 24 cores on the PBMC data.
}
\examples{
\dontrun{
write_sparse_matrix_from_bam("pbmc_data/possorted_bam.bam", "chrom1.mm", region = gintervals.all()[1, ], cell_names = atac_sc)
}

}
